% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{add_delta_report_bimestre_coesione}
\alias{add_delta_report_bimestre_coesione}
\title{Integra report bimestre con bimestre precedente a scelta}
\usage{
add_delta_report_bimestre_coesione(report, path_to_old, export = FALSE)
}
\arguments{
\item{report}{Report da make_report_bimestre_coesione.}

\item{path_to_old}{Percorso a elaborazione per bimestre precedente fino a versione, senza nome file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per ciclo e ambito.
}
\description{
Integra report bimestre con bimestre precedente a scelta.
}
