% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{add_old_turismo}
\alias{add_old_turismo}
\title{Addendum Turismo}
\usage{
add_old_turismo(pseudo, export = TRUE, debug = FALSE)
}
\arguments{
\item{pseudo}{Dataset "pseudo.csv"}
}
\value{
Un dataframe pseudo.
}
\description{
....
}
\section{Warning}{

Punta a "old_perim.csv".
Forse non serve nemmeno più...
}

