% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{analisi_data_quality_psc}
\alias{analisi_data_quality_psc}
\title{Analisi qualità dati PSC}
\usage{
analisi_data_quality_psc(
  progetti_psc,
  interventi_psc,
  progetti,
  export = FALSE
)
}
\arguments{
\item{progetti_psc}{Dataset da load_progetti_psc_migrati()}

\item{interventi_psc}{Dataset da DBCOE per programmazione interventi PSC}

\item{progetti}{Dataset progetti da load_progetti(visualizzati = FALSE, light = TRUE)}

\item{export}{Vuoi esportare il file in OUTPUT?}
}
\value{
Dataframe
}
\description{
Analisi qualità dati PSC
}
