% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_asset.R
\docType{data}
\name{aree_temi_fsc}
\alias{aree_temi_fsc}
\title{Elenco aree tematiche e temi prioritari FSC 2014-2020}
\format{
A data frame
}
\source{
Team OC
}
\usage{
aree_temi_fsc
}
\description{
Input per la funzione ... (TODO).
}
\details{
Le variabili sono:
\itemize{
  \item x_CICLO: ...
  \item \strong{COD_SETTORE_STRATEGICO}: area tematica FSC (denominazione BDU?)
  \item DESCR_SETTORE_STRATEGICO_FSC: denominazione area tematica FSC (denominazione BDU?)
  \item x_AREA_TEMATICA: denominazione area tematica FSC
  \item \strong{COD_ASSE_TEMATICO}: tema prioritario FSC (denominazione BDU?)
  \item DESCR_ASSE_TEMATICO_FSC: denominazione tema prioritario FSC (denominazione BDU?)
  \item x_TEMA_PRIORITARIO: denominazione tema prioritario FSC
  \item QUERY: variabile guida per query (inserire 1, 2 o 0)
  \item NOTE: ...
}
}
\section{Warning}{

Da integrare con RA rural (CHK?).
}

\keyword{datasets}
