% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_asset.R
\docType{data}
\name{categorie_ue}
\alias{categorie_ue}
\title{Elenco campi di intervento UE}
\format{
A data frame
}
\source{
Team OC su dati BDU
}
\usage{
categorie_ue
}
\description{
Input per la funzione query_ue().
Creata con make_matrix_ue() in setup_data.R.
Comprende i temi prioritari 2007-2013 e i campi di intervento 2014-2020.
}
\details{
Le variabili sono:
\itemize{
  \item OC_COD_CICLO: ...
  \item \strong{COD_TEMA_CAMPO}: codice per tema 2007-2013 o campo d'intervento 2014-2020 (con leading zero)
  \item DESCR_TEMA_CAMPO: ...
  \item QUERY: variabile guida per query (inserire 1, 2 o 0)
  \item NOTE: ...
}

#' @section Warning:
Dati da aggiornare per ogni bimestre.
Comprende assegnazioni multiple possibili in BDU 2014-2020 (CHK?)
}
\keyword{datasets}
