% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{chk_allineamento_costo_coe}
\alias{chk_allineamento_costo_coe}
\title{Varifica totale di COE}
\usage{
chk_allineamento_costo_coe(
  report,
  programmi,
  report_macroaree,
  perimetro = NULL
)
}
\arguments{
\item{report}{Dataframe da make_report_bimestre_coesione()}

\item{programmi}{Dataframe da make_report_programmi_coesione()}

\item{report_macroaree}{Dataframe da make_report_macroaree_coesione()}

\item{perimetro}{Dataframe da prep_perimetro_bimestre_coesione().E' opzionale}
}
\value{
Dataframe per ciclo e ambito con confronto del totale di RISORSE in diversi report di octk.
}
\description{
Report con apertura ciclo e ambito con confronto del totale di COE in diversi report di octk.
}
