% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_allineamento_dati_info}
\alias{chk_allineamento_dati_info}
\title{Verifica allineamento dei dati finanziari storicizzati nel DBCOE}
\usage{
chk_allineamento_dati_info(export = FALSE)
}
\arguments{
\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica allineamento nel DBCOE confrontando file "Dati" e file "Info".
}
