% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_allineamento_interventi_accordi}
\alias{chk_allineamento_interventi_accordi}
\title{Verifica allineamento Accordi nel DBCOE}
\usage{
chk_allineamento_interventi_accordi(export = FALSE)
}
\arguments{
\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica allineamento Accordi nel DBCOE confrontando file "Dati" e file "Interventi".
}
