% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_allineamento_interventi_psc}
\alias{chk_allineamento_interventi_psc}
\title{Verifica allineamento PSC nel DBCOE}
\usage{
chk_allineamento_interventi_psc(export = FALSE)
}
\arguments{
\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica allineamento della sezione ordinaria dei PSC nel DBCOE confrontando file "Dati" e file "Interventi".
}
