% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{chk_allineamento_risorse}
\alias{chk_allineamento_risorse}
\title{Varifica totale di RISORSE}
\usage{
chk_allineamento_risorse(report, programmi, report_macroaree, risorse = NULL)
}
\arguments{
\item{report}{Dataframe da make_report_bimestre_coesione()}

\item{programmi}{Dataframe da make_report_programmi_coesione()}

\item{report_macroaree}{Dataframe da make_report_macroaree_coesione()}

\item{risorse}{Dataframe da make_report_risorse().E' opzionale}
}
\value{
Dataframe per ciclo e ambito con confronto del totale di RISORSE in diversi report di octk.
}
\description{
Report con apertura ciclo e ambito con confronto del totale di RISORSE in diversi report di octk.
}
