% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_allineamento_risorse_ue_siepoc1420}
\alias{chk_allineamento_risorse_ue_siepoc1420}
\title{Verifica allineamento risorse UE 1420 nel DBCOE}
\usage{
chk_allineamento_risorse_ue_siepoc1420(export = FALSE)
}
\arguments{
\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica allineamento della risorse UE 1420 nel DBCOE confrontando file "Dati", "Correzioni" e "Stime".
}
