% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{chk_bimestre_light}
\alias{chk_bimestre_light}
\title{Controlla integrità file light}
\usage{
chk_bimestre_light(progetti, operazioni)
}
\arguments{
\item{progetti}{Dataset in formato standard da load_progetti().}

\item{operazioni}{Dataset in formato standard da load_progetti().}
}
\value{
Messaggi e report di controllo.
}
\description{
Controlla integrità file light
}
