% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{chk_bimestre_sub_po_riclass}
\alias{chk_bimestre_sub_po_riclass}
\title{Workflow chk bimestre - Controllo x_vars}
\usage{
chk_bimestre_sub_po_riclass(progetti, export = FALSE)
}
\arguments{
\item{progetti}{File di tipo progetti da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{export}{Vuoi esportare i report di controllo in TEMP?}
}
\value{
Messaggi e report di controllo.
}
\description{
Workflow di controllo dei dati del nuovo bimestre - Controllo su po_riclass.csv e x_vars
}
