% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{chk_dati_dbcoe_psc_migrati}
\alias{chk_dati_dbcoe_psc_migrati}
\title{Verifica coerenza tra file dati FSC e file interventi PSC}
\usage{
chk_dati_dbcoe_psc_migrati()
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{versione_psc}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}
}
\value{
Dataframe
}
\description{
Verifica coerenza tra file dati FSC e file interventi PSC
}
