% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{chk_delta_db_pnrr}
\alias{chk_delta_db_pnrr}
\title{Confronta due versioni del database PNRR}
\usage{
chk_delta_db_pnrr(DB_new, DB_old)
}
\arguments{
\item{DB_new}{Percorso al database più recente}

\item{DB_old}{Percorso al database precedente}
}
\value{
Lista con i dataframe delle differenze
}
\description{
Confronta due versioni del database PNRR identificando misure nuove, cancellate e con importi modificati
}
