% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{chk_delta_psc_migrati_bimestre}
\alias{chk_delta_psc_migrati_bimestre}
\title{Verifica variazioni dati FSC per elaborazioni PSC}
\usage{
chk_delta_psc_migrati_bimestre(
  bimestre,
  versione_psc,
  bimestre_old,
  versione_psc_old
)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{versione_psc}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}

\item{bimestre_old}{Bimestre di confronto}

\item{versione_psc_old}{Versione di confronto dei dati  (sono possibili più versioni per lo stesso bimestre)}
}
\value{
File "dati_psc_BIMESTRE_VERSIONE.csv" in OUTPUT
}
\description{
Verifica variazioni dati FSC per elaborazioni PSC rispetto a un bimestre e versione precedente
}
\note{
...
}
