% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{chk_delta_scarico}
\alias{chk_delta_scarico}
\title{Confronta versioni dati PNRR}
\usage{
chk_delta_scarico(
  progetti_pnrr,
  bimestre_pnrr_old,
  versione_pnrr_old,
  export = TRUE
)
}
\arguments{
\item{progetti_pnrr}{Dataframe dei progetti da REGIS di riferimento (output di setup_dati_pnrr())}

\item{bimestre_pnrr_old}{Bimestre di confronto per i dati PNRR}

\item{versione_pnrr_old}{Versione di confronto dei dati (sono possibili più versioni per lo stesso bimestre)}

\item{export}{Vuoi esportare le misure mancanti in Excel? Default TRUE}

\item{pagamenti}{Dataframe dei pagamenti dei progetti da REGIS (output di load_pagamenti_regis)}
}
\value{
Dataframe con dati di attuazione PNRR
}
\description{
Utility di confronta tra la versione attuale dei dati PNRR e una precedente
}
