% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{chk_risorse_ciclo_contabile_strategia}
\alias{chk_risorse_ciclo_contabile_strategia}
\title{Verifica risorse per ciclo contabile e ciclo strategia}
\usage{
chk_risorse_ciclo_contabile_strategia(
  use_flt = TRUE,
  force_yei = FALSE,
  use_eu = FALSE,
  export = TRUE
)
}
\arguments{
\item{export}{vuoi salvare il file?}

\item{programmi_new}{Dataframe da make_report_programmi_coesione()}

\item{programmi_old}{Dataframe da make_report_programmi_coesione()}

\item{path_to_new}{Percorso a csv generato con make_report_programmi_coesione().}

\item{path_to_old}{Percorso a csv generato con make_report_programmi_coesione().}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica risorse per ciclo contabile e ciclo strategia.
}
