% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_totali_fsc}
\alias{chk_totali_fsc}
\title{Verifica allineamento totali finanziari FSC}
\usage{
chk_totali_fsc()
}
\value{
Un messaggio di log.
}
\description{
Verifica allineamento del file "dati" con i totali finanziari FSC
}
