% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_variazione_dbcoe}
\alias{chk_variazione_dbcoe}
\title{Verifica variazione risorse per programma nel DBCOE}
\usage{
chk_variazione_dbcoe(
  dbcoe_new = NULL,
  dbcoe_old = NULL,
  use_cicli_psc = FALSE,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE,
  export = FALSE
)
}
\arguments{
\item{dbcoe_new}{Versione attuale del DBCOE. Di default è quella configurata in oc_init().}

\item{dbcoe_old}{Versione precedente del DBCOE.}

\item{use_cicli_psc}{Vuoi usare i dati di programmazione per cicli dei PSC?}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni di decisioni in base alle delibere sui POC?}

\item{stime_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le stime di chiusura dei programmi?}

\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica variazione risorse per programma nel DBCOE. Confronta due versioni del database.
}
