% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_report_coesione.R
\name{chk_variazione_programmi_coesione}
\alias{chk_variazione_programmi_coesione}
\title{Verifica variazione variabili coesione per programma}
\usage{
chk_variazione_programmi_coesione(
  programmi = NULL,
  dati_new = NULL,
  dbcoe_new = NULL,
  dati_old,
  dbcoe_old,
  usa_meuro = FALSE,
  show_cp = FALSE,
  use_eu = FALSE,
  use_cicli_psc = FALSE,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE,
  use_flt = TRUE,
  use_pqt = FALSE,
  export = FALSE,
  export_sum = FALSE
)
}
\arguments{
\item{programmi}{Report da make_report_programmi_coesione()}

\item{dati_new}{Versione attuale dei dati. Di default è quella configurata in oc_init(), coincide con "bimestre".}

\item{dbcoe_new}{Versione attuale del DBCOE. Di default è quella configurata in oc_init().}

\item{dati_old}{Versione precedente dei dati (espressa come bimestre).}

\item{dbcoe_old}{Versione precedente del DBCOE.}

\item{use_cicli_psc}{Vuoi usare i dati di programmazione per cicli dei PSC?}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni di decisioni in base alle delibere sui POC?}

\item{stime_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le stime di chiusura dei programmi?}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{export}{Vuoi salvare il file?}

\item{export_sum}{Vuoi salvare il file con la sintesi per ciclo/ambito?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica variazione variabili coesione per programma. Confronta RISORSE per due versioni del DBCOE e COE, COE_IMP e COE_PAG pert due bimestri.
}
