% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{chk_variazione_risorse_ciclo_ambito}
\alias{chk_variazione_risorse_ciclo_ambito}
\title{Verifica variazione risorse per ciclo e ambito nei report per la pubblicazione}
\usage{
chk_variazione_risorse_ciclo_ambito(
  risorse_new = NULL,
  risorse_old = NULL,
  path_to_new = NULL,
  path_to_old = NULL,
  export = FALSE
)
}
\arguments{
\item{risorse_new}{Dataframe da make_report_risorse()}

\item{risorse_old}{Dataframe da make_report_risorse()}

\item{path_to_new}{Percorso ad attuale folder in cui si trovano i file "risorse_coesione_2014-2020.csv" e "risorse_coesione_2007-2013.csv" generati con make_report_risorse().}

\item{path_to_old}{Percorso a precedente folder in cui si trova il file "risorse_coesione.csv" oppure i file "risorse_coesione_2014-2020.csv" e "risorse_coesione_2007-2013.csv" generati con make_report_risorse().}

\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per ciclo e ambito.
}
\description{
Verifica variazione risorse per ciclo e ambito nei report per la pubblicazione. Confronta due dataframe risultanti da make_report_risorse() o i file csv da questa esportati.
}
