% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataiku.R
\name{chk_variazione_sintesi_coesione_dataiku}
\alias{chk_variazione_sintesi_coesione_dataiku}
\title{Verifica variazione variabili coesione per programma. Confronta RISORSE per due versioni del DBCOE e COE, COE_IMP e COE_PAG pert due bimestri.}
\usage{
chk_variazione_sintesi_coesione_dataiku(
  dati_new,
  dbcoe_new,
  dati_old,
  dbcoe_old,
  use_cicli_psc = FALSE,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE,
  use_flt = TRUE,
  export = FALSE
)
}
\arguments{
\item{dati_new}{Versione attuale dei dati. Di default è quella configurata in oc_init(), coincide con "bimestre".}

\item{dbcoe_new}{Versione attuale del DBCOE. Di default è quella configurata in oc_init().}

\item{dati_old}{Versione precedente dei dati (espressa come bimestre).}

\item{dbcoe_old}{Versione precedente del DBCOE.}

\item{use_cicli_psc}{Vuoi usare i dati di programmazione per cicli dei PSC?}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni di decisioni in base alle delibere sui POC?}

\item{stime_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le stime di chiusura dei programmi?}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica variazione variabili coesione per programma. Confronta RISORSE per due versioni del DBCOE e COE, COE_IMP e COE_PAG pert due bimestri.
}
