% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_accordi.R
\name{chk_variazioni_accordi}
\alias{chk_variazioni_accordi}
\title{Verifica variazioni interventi accordi}
\usage{
chk_variazioni_accordi(db_ver_new, db_ver_old)
}
\arguments{
\item{db_ver_new}{Nuova versione del DBCOE da confrontare.}

\item{db_ver_old}{Precedente versione del DBCOE da confrontare.}
}
\value{
Salva in TEMP 3 report in ofrmalto long (analisi variazioni, sintesi per regione e sintesi totali)
}
\description{
Verifica variazioni tra due versioni del database interventi degli accordi
}
