% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{clp_perimetro}
\alias{clp_perimetro}
\title{Assestare lista finale perimetro}
\usage{
clp_perimetro(progetti, perimetro, export = FALSE)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}

\item{perimetro}{definire il dataset del perimetro finale che contiene solo i CLP}

\item{export}{settare a TRUE per scaricare CSV finale in OUTPUT}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO.
}
\description{
Assestare lista finale perimetro per consentire il caricamento con CUP aggregati
}
