% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{convert_input_csv_to_xls}
\alias{convert_input_csv_to_xls}
\title{Converte input da csv a excel}
\usage{
convert_input_csv_to_xls(OLD)
}
\arguments{
\item{OLD}{Percorso alla cartella di livello WORK di una specifica elaborazione, anche diversa dal bimestre di riferimento.}
}
\value{
Prende tutti i csv nel folder indicato e li inserisce in un file excel in format query_input.xlsx
}
\description{
Converte gli input in formato csv, ora deprecati, nel nuovo format su file excel.
}
