% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_accordi.R
\name{crea_accordi_pubblicazione_campania}
\alias{crea_accordi_pubblicazione_campania}
\title{Crea template pubblicazione accordi per regione Campania}
\usage{
crea_accordi_pubblicazione_campania(regione, interventi)
}
\arguments{
\item{regione}{valore per selezionare AMMINISTRAZIONE_TITOLARE}

\item{interventi}{Elenco interventi da load_db_accordi()}
}
\value{
File in output nella cartella "pubblicazione"
}
\description{
Crea template pubblicazione accordi per una regione - Versione speciale per Campania
}
