% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_accordi.R
\name{crea_report_accordi_monitoraggio}
\alias{crea_report_accordi_monitoraggio}
\title{Crea template monitoraggio extra-sistema accordi per una regione}
\usage{
crea_report_accordi_monitoraggio(regione, interventi, template)
}
\arguments{
\item{regione}{valore per selezionare AMMINISTRAZIONE_TITOLARE}

\item{interventi}{Elenco interventi da load_db_accordi()}

\item{template}{Nome del file xlsx in input}
}
\value{
File in output nella cartella "monitoraggio"
}
\description{
Crea template monitoraggio extra-sistema accordi per una regione
}
