% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cd.R
\name{export_report_cd_regioni}
\alias{export_report_cd_regioni}
\title{Esporta report CD in xlsx}
\usage{
export_report_cd_regioni(
  report,
  titolo = NULL,
  focus = "elab",
  ciclo = NULL,
  ambito = NULL,
  gruppo = NULL
)
}
\arguments{
\item{report}{dataframe da report_cd_regioni().}

\item{titolo}{titolo interno al report (chr).}

\item{focus}{nome per file.}

\item{ambito}{...}

\item{gruppo}{...}

\item{...}{...}
}
\value{
Un file xlsx con apertura per regione e fase procedurale.
}
\description{
Esporta report CD in xlsx da template "stato_template_regioni.xlsx"
}
