% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_peri.R
\name{export_report_edit}
\alias{export_report_edit}
\title{Wrapper per report con lista di contenuti editabile}
\usage{
export_report_edit(
  perimetro,
  report_ls,
  nome_file = NULL,
  debug = FALSE,
  use_template = FALSE
)
}
\arguments{
\item{perimetro}{Dataset da workflow_report}

\item{report_ls}{Elenco dei report da generare}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_[1], QUERY_[2], QUERY_[N] e TIPO_QUERY.
}
\description{
Wrapper per report con lista di contenuti editabile.
}
