% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_regioni.R
\name{export_report_regione}
\alias{export_report_regione}
\title{Esporta report regionale in Excel}
\usage{
export_report_regione(report, regione, bimestre = bimestre,
  OUTPUT = OUTPUT)
}
\arguments{
\item{report}{Dataset prodotto con report_regione().}

\item{regione}{Regione di interesse.}

\item{bimestre}{Bimestre di riferimento in formato OC.}
}
\value{
Il dataset con il report regionale in Excel.
}
\description{
Esporta report regionale in Excel
}
