% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{fix_assegnazioni_legge_pra_campania}
\alias{fix_assegnazioni_legge_pra_campania}
\title{Cambia programma per progetti in assegnazioni di legge PRA Campania}
\usage{
fix_assegnazioni_legge_pra_campania(progetti_psc, progetti)
}
\arguments{
\item{progetti_psc}{Dataset progetti PSC da dentro prep_dati_psc_bimestre()}

\item{progetti}{Dataset progetti da load_progetti(bimestre, visualizzati = FALSE, light = FALSE), come usato in dentro prep_dati_psc_bimestre()}
}
\value{
dataframe
}
\description{
Cambia programma per progetti in assegnazioni di legge PRA Campania, che sono ricompresi nelle articolazioni 21 (rifiuti) e 22 (compensazioni ambientali).
}
\note{
Da usare dentro prep_dati_psc_bimestre(). Deve seguire fix_visualizzati_pra_campania() altrimenti il fix visualizzati non funziona perché qui cambia il codice programma
}
