% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{fix_extra_2127}
\alias{fix_extra_2127}
\title{Fix temporaneo per i dataset operazioni}
\usage{
fix_extra_2127(df)
}
\arguments{
\item{df}{Dataset in formato standard.}
}
\value{
Il dataset operazioni integrato.
}
\description{
Integra il dataset.
}
