% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{fix_id_psc_15_digit}
\alias{fix_id_psc_15_digit}
\title{Correzione per codici PSC superiori a 15 digit}
\usage{
fix_id_psc_15_digit(df, var1 = "ID_PSC")
}
\arguments{
\item{df}{Dataframe}

\item{var1}{Nome variabile da modificare (OC_CODICE_PROGRAMMA oppure ID_PSC)}
}
\value{
Dataframe con valori modificati
}
\description{
Correzione per codici PSC superiori a 15 digit
}
