% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{fix_id_psc_ministeri}
\alias{fix_id_psc_ministeri}
\title{Correzione per denominazioni PSC ministeri}
\usage{
fix_id_psc_ministeri(df, var1 = "PSC")
}
\arguments{
\item{df}{Dataframe con variabili}

\item{var1}{Nome variabile da modificare (PSC oppure DESCRIZIONE_PROGRAMMA)}
}
\value{
Dataframe con valori modificati per la variabile PSC oppure DESCRIZIONE_PROGRAMMA in funzione dei valori diu ID_PSC oppure OC_CODICE_PROGRAMMA
}
\description{
Correzione per denominazioni PSC ministeri
}
