% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{fix_macroaree_713}
\alias{fix_macroaree_713}
\title{Fix temporaneo per i dataset operazioni (versione per 713)}
\usage{
fix_macroaree_713(df)
}
\arguments{
\item{df}{Dataset in formato standard.}
}
\value{
Il dataset operazioni integrato.
}
\description{
Integra il dataset risolvendo il problema di caratteri spuri. Attenzione perché va fatto anche in "progetti".
}
