% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{fix_macroaree_progetti_sie}
\alias{fix_macroaree_progetti_sie}
\title{Corregge progetti con macroaree da livelli gerarchici in operazioni}
\usage{
fix_macroaree_progetti_sie(bimestre, progetti, operazioni, export = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{progetti}{Dataset di classe "progetti".}

\item{operazioni}{Dataset di classe "operazioni"}

\item{export}{vuoi salvare il file?}
}
\value{
Dataset macroaree per i report make_report_programmi_macroaree e make_report_macroaree_coesione
}
\description{
Corregge progetti con macroaree da livelli gerarchici in operazioni
}
