% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{fix_operazioni_713_macroaree}
\alias{fix_operazioni_713_macroaree}
\title{Fix temporaneo per i dataset operazioni (versione per 713)}
\usage{
fix_operazioni_713_macroaree(df)
}
\arguments{
\item{df}{Dataset in formato standard.}
}
\value{
Il dataset operazioni integrato.
}
\description{
Integra il dataset risolvendo il problema di caratteri spuri. Attenzione perché va fatto anche in "progetti".
}
