% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataiku.R
\name{fix_operazioni_dataiku}
\alias{fix_operazioni_dataiku}
\title{Fix temporaneo per i dataset operazioni}
\usage{
fix_operazioni_dataiku(df)
}
\arguments{
\item{df}{Dataset in formato standard.}
}
\value{
Il dataset operazioni integrato.
}
\description{
Integra il dataset.
}
