% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{fix_operazioni_macroaree}
\alias{fix_operazioni_macroaree}
\title{Fix temporaneo per i dataset operazioni}
\usage{
fix_operazioni_macroaree(df)
}
\arguments{
\item{df}{Dataset in formato standard.}
}
\value{
Il dataset operazioni integrato.
}
\description{
Integra il dataset.
}
