% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_progetti.R
\name{fix_progetti}
\alias{fix_progetti}
\title{Fix temporaneo per il dataset progetti in vestione preesteso}
\usage{
fix_progetti(progetti)
}
\arguments{
\item{progetti}{Dataset in formato standard.}
}
\value{
Il dataset progetti integrato.
}
\description{
Integra il dataset.
}
