% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{fix_progetti_sgp}
\alias{fix_progetti_sgp}
\title{#' Calcola costo realizzato
#'
#' Calcola costo realizzato con riproporzionamento su risorse coesione
#' 
#' @param bimestre Bimestre di riferimento
#' @param filename Nome file xlsx in DATI > SGP
#' @param chk_today Parametro da passare a get_stato_attuazione(), con formato "2021-02-28"
#' @param matrix_po_psc Matrice di riconciliazione PO - PSC
#' @return File "dati_sgp_BIMESTRE.csv" in TEMP 
#' @note ...
get_costo_realizzato_713_1420 <- function(progetti_psc, progetti, progetti_sgp, matrix_po_psc) {
  
  # clean sgp
  progetti_sgp <- progetti_sgp %>% 
    select(COD_LOCALE_PROGETTO, TITOLO, 
           DENOMINAZIONE_INTESA, CODICE_STRUMENTO,
           FINANZIAMENTO_FSC_NETTO, FINANZIAMENTO_TOTALE_PUBBLICO_NETTO, 
           TOTALE_FINANZIAMENTI_PVT, TOTALE_ECONOMIE_PVT, 
           TOTALE_FINANZIAMENTI, TOTALE_ECONOMIE,
           IMPORTO_NAZIONALE, IMPORTO_REGIONALE, IMPORTO_NON_DEFINITO,
           COSTO_REALIZZATO, COSTO_REALIZZATO_1,
           IMPEGNI,
           PAGAMENTI_TOTALI, PAGAMENTI_FSC) %>% 
    mutate(OC_CODICE_PROGRAMMA = case_when(DENOMINAZIONE_INTESA == "ABRUZZO" ~ "33",
                                           DENOMINAZIONE_INTESA == "BASILICATA" ~ "37",
                                           DENOMINAZIONE_INTESA == "CALABRIA" ~ "38",
                                           DENOMINAZIONE_INTESA == "CAMPANIA" ~ "35",
                                           DENOMINAZIONE_INTESA == "EMILIA-ROMAGNA" ~ "28",
                                           DENOMINAZIONE_INTESA == "FRIULI-VENEZIA GIULIA" ~ "26",
                                           DENOMINAZIONE_INTESA == "LAZIO" ~ "32",
                                           DENOMINAZIONE_INTESA == "LIGURIA" ~ "27",
                                           DENOMINAZIONE_INTESA == "LOMBARDIA" ~ "01",
                                           DENOMINAZIONE_INTESA == "MARCHE" ~ "04",
                                           DENOMINAZIONE_INTESA == "MOLISE" ~ "34",
                                           DENOMINAZIONE_INTESA == "P.A. BOLZANO" ~ "41",
                                           DENOMINAZIONE_INTESA == "P.A. TRENTO" ~ "40",
                                           DENOMINAZIONE_INTESA == "PIEMONTE" ~ "21",
                                           DENOMINAZIONE_INTESA == "PUGLIA" ~ "36",
                                           DENOMINAZIONE_INTESA == "SARDEGNA" ~ "05",
                                           DENOMINAZIONE_INTESA == "SICILIA" ~ "39",
                                           DENOMINAZIONE_INTESA == "TOSCANA" ~ "02",
                                           DENOMINAZIONE_INTESA == "UMBRIA" ~ "03",
                                           DENOMINAZIONE_INTESA == "VALLE D'AOSTA" ~ "22",
                                           DENOMINAZIONE_INTESA == "VENETO" ~ "25"))
  
  
  # calcola quota 713-1420
  message("calcola cr per 713 e 1420")
  appo <- progetti_psc %>% 
    # filter(OC_FLAG_VISUALIZZAZIONE == 0) %>% 
    left_join(progetti %>% 
                select(COD_LOCALE_PROGETTO, COSTO_REALIZZATO, OC_FINANZ_STATO_FSC_NETTO, OC_FINANZ_TOT_PUB_NETTO, OC_FINANZ_PRIVATO_NETTO),
              by = "COD_LOCALE_PROGETTO") %>% 
    mutate_if(is.numeric, replace_na, replace = 0) %>% 
    mutate(FINANZ_TOT = OC_FINANZ_TOT_PUB_NETTO + OC_FINANZ_PRIVATO_NETTO,
           # x = OC_FINANZ_STATO_FSC_NETTO/FINANZ_TOT,
           x = COE/FINANZ_TOT,
           COSTO_REALIZZATO_2 = COSTO_REALIZZATO * x)
  
  dim(appo)[1] == dim(progetti_psc)[1]
  
  
  appo %>% 
    summarise(COE = sum(COE, na.rm = TRUE),
              COE_PAG = sum(COE_PAG, na.rm = TRUE),
              COSTO_REALIZZATO = sum(COSTO_REALIZZATO, na.rm = TRUE),
              OC_FINANZ_STATO_FSC_NETTO = sum(OC_FINANZ_STATO_FSC_NETTO, na.rm = TRUE),
              OC_FINANZ_TOT_PUB_NETTO = sum(OC_FINANZ_TOT_PUB_NETTO, na.rm = TRUE),
              COSTO_REALIZZATO_2 = sum(COSTO_REALIZZATO_2, na.rm = TRUE))
  # A tibble: 1 x 6
  # COE      COE_PAG COSTO_REALIZZATO OC_FINANZ_STATO_FSC_NETTO OC_FINANZ_TOT_PUB_NETTO COSTO_REALIZZATO_2
  # <dbl>        <dbl>            <dbl>                     <dbl>                   <dbl>              <dbl>
  # 49290164401. 10040154286.     21986749485.              50629969321.            68164313459.       15828065459.
  
  # report per programma
  # appo1 <- appo %>% 
  #   group_by(ID_PSC, PSC, OC_CODICE_PROGRAMMA, x_PROGRAMMA, x_CICLO) %>% 
  #   summarise(COE = sum(COE, na.rm = TRUE),
  #             COE_IMP = sum(COE_IMP, na.rm = TRUE),
  #             COE_PAG = sum(COE_PAG, na.rm = TRUE),
  #             COSTO_REALIZZATO = sum(COSTO_REALIZZATO_2, na.rm = TRUE))
  
  # integra 06
  message("calcola cr per 06")
  temp <- progetti_sgp %>% 
    mutate(x_CICLO = "2000-2006",
           x_PROGRAMMA = paste0("INTESA ", DENOMINAZIONE_INTESA)) %>% 
    left_join(matrix_po_psc %>% 
                select(OC_CODICE_PROGRAMMA, ID_PSC, PSC),
              by = "OC_CODICE_PROGRAMMA") %>% 
    mutate_if(is.numeric, replace_na, replace = 0) %>% 
    mutate(FIN_TOT = TOTALE_FINANZIAMENTI - TOTALE_ECONOMIE,
           x = FINANZIAMENTO_FSC_NETTO / FIN_TOT,
           k = IMPORTO_REGIONALE / (IMPORTO_NAZIONALE + IMPORTO_REGIONALE + IMPORTO_NON_DEFINITO),
           COSTO_REALIZZATO_2 = COSTO_REALIZZATO * x * k,
           COE = FINANZIAMENTO_FSC_NETTO * k,
           COE_IMP = IMPEGNI * x * k,
           COE_PAG = PAGAMENTI_TOTALI * x * k)
  
  dim(temp)[1] == dim(progetti_sgp)[1]
  
  # temp %>% 
  #   summarise(COE = sum(COE, na.rm = TRUE),
  #             PAGAMENTI_TOTALI = sum(PAGAMENTI_TOTALI, na.rm = TRUE),
  #             PAGAMENTI_FSC = sum(PAGAMENTI_FSC, na.rm = TRUE),
  #             COSTO_REALIZZATO = sum(COSTO_REALIZZATO, na.rm = TRUE),
  #             FINANZIAMENTO_FSC_NETTO = sum(FINANZIAMENTO_FSC_NETTO, na.rm = TRUE),
  #             FINANZIAMENTO_TOTALE_PUBBLICO_NETTO = sum(FINANZIAMENTO_TOTALE_PUBBLICO_NETTO, na.rm = TRUE),
  #             IMPORTO_REGIONALE = sum(IMPORTO_REGIONALE, na.rm = TRUE),
  #             IMPORTO_NAZIONALE = sum(IMPORTO_NAZIONALE, na.rm = TRUE),
  #             IMPORTO_NON_DEFINITO = sum(IMPORTO_NON_DEFINITO, na.rm = TRUE),
  #             COSTO_REALIZZATO_2 = sum(COSTO_REALIZZATO_2, na.rm = TRUE))
  # temp %>% 
  #   summarise(COE = sum(COE, na.rm = TRUE),
  #             COE_IMP = sum(COE_IMP, na.rm = TRUE),
  #             COE_PAG = sum(COE_PAG, na.rm = TRUE),
  #             COSTO_REALIZZATO_2 = sum(COSTO_REALIZZATO_2, na.rm = TRUE))
  # 
  # temp1 <- temp %>% 
  #   group_by(ID_PSC, PSC, OC_CODICE_PROGRAMMA, x_PROGRAMMA, x_CICLO) %>% 
  #   summarise(COE = sum(COE, na.rm = TRUE),
  #             COE_IMP = sum(COE_IMP, na.rm = TRUE),
  #             COE_PAG = sum(COE_PAG, na.rm = TRUE),
  #             COSTO_REALIZZATO = sum(COSTO_REALIZZATO_2, na.rm = TRUE))
  
  # bind
  # appo2 <- appo1 %>%
  #   bind_rows(temp1)
  message("bind di 713-1420 e 06")
  
  appo2 <- appo %>%
    filter(x_CICLO != "2000-2006") %>% 
    bind_rows(temp) %>% 
    select(COD_LOCALE_PROGETTO, COSTO_REALIZZATO_2) %>% 
    rename(COSTO_REALIZZATO = COSTO_REALIZZATO_2)
  
  chk <- appo %>% 
    bind_rows(temp) %>% 
    mutate(CHK = COSTO_REALIZZATO_2 - COE) %>% 
    filter(CHK > 1)
  
  chk %>% count(x_CICLO)
  # CHK: queste sono le uniche anomalie?
  
  # export
  message("export")
  dim(progetti_psc)[1] == dim(appo2)[1]
  
  progetti_psc <- progetti_psc %>% 
    left_join(appo2, by = "COD_LOCALE_PROGETTO")
  
  return(progetti_psc)
  
}}
\usage{
fix_progetti_sgp(progetti_sgp, bimestre)
}
\arguments{
\item{progetti_sgp}{Dataset progetti PSC da dentro prep_dati_sgp_bimestre()}

\item{bimestre}{Bimestre di riferimento}
}
\value{
dataframe
}
\description{
Correzioni dati SGP
}
\details{
Correzioni dati SGP per bimestre
}
\note{
Da usare dentro prep_dati_sgp_bimestre()
}
