% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{fix_snai}
\alias{fix_snai}
\title{Fix variabili SNAI}
\usage{
fix_snai(progetti, path_snai)
}
\arguments{
\item{progetti}{Dataset in formato standard.}
}
\value{
Il dataset progetti integrato.
}
\description{
Fix temporaneo per integrare le variabili SNAI (COD_AREA_INT e AREA_INTERNA) dal file di Andrea
}
