% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{fix_visualizzati_apq_mari_marche}
\alias{fix_visualizzati_apq_mari_marche}
\title{Correzione progetti con OC_FLAG_VISUALIZZAZIONE = 5 in APQ MARI Marche}
\usage{
fix_visualizzati_apq_mari_marche(progetti_psc)
}
\arguments{
\item{progetti_psc}{Dataset progetti PSC da dentro prep_dati_psc_bimestre()}
}
\value{
dataframe
}
\description{
Correzione progetti con OC_FLAG_VISUALIZZAZIONE = 5 in APQ MARI Marche
}
\note{
Da usare dentro prep_dati_psc_bimestre()
}
