% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{get_catreg_UE}
\alias{get_catreg_UE}
\title{Definisce la categoria di regioni UE}
\usage{
get_catreg_UE(df, progetti = NULL, real_reg = TRUE, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti" originale per integrazione.}

\item{real_reg}{Vuoi usare le regioni reali?.}

\item{debug_mode}{Dataset in formato "progetti".}
}
\value{
Il dataset con la variabile x_CATREG, come factor con levels = c("RS", "RT", "RMS", "Trasversale", "Nazionale", "Estero").
}
\description{
Crea la variabile x_CATREG con la denominazione della Regione o della Provincia Autonoma.
Le altre denominazioni sono semplificate in "ALTRO TERRITORIO".
}
\note{
La modalità **debug** non + implementata. La modalità **real_reg** forza tutti i progetti
di un programma regionale sul territorio della Regione e sulla relativa macroarea.
}
