% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{get_dimensione_fin}
\alias{get_dimensione_fin}
\title{Integra classi di dimensione finanziaria}
\usage{
get_dimensione_fin(df, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}
}
\value{
Il dataset con la variabile CLASSE_FIN, come factor con levels = c("0-100k", "100k-500k", "500k-1M", "1M-2M", "2M-5M", "5M-10M", "10M-infty")
}
\description{
Calcola e integra una variabile (CLASSE_FIN) con le classi di dimensione finanziaria.
}
\note{
La modalità **debug** non + implementata.
}
