% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{get_macroarea_localizzazioni}
\alias{get_macroarea_localizzazioni}
\title{Integra la macroarea da localizzazioni}
\usage{
get_macroarea_localizzazioni(df, progetti, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti" in formato PREESTESO per integrazione.}

\item{debug_mode}{Dataset in formato "progetti".}
}
\value{
Il dataset con la variabile x_MACROAREA, come factor con levels = c("Centro-Nord", "Sud", "Trasversale", "Nazionale", "Estero").
}
\description{
Integra in x_MACROAREA la macroarea dalle localizzazioni di PREESTESO partendo da OC_MACROAREA.
}
