% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{get_real_reg}
\alias{get_real_reg}
\title{Definisce la regione reale}
\usage{
get_real_reg(df, progetti, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti".}
}
\value{
Il dataset con la variabile COD_REGIONE modificata.
}
\description{
Sovrascrive la variabile COD_REGIONE forzando tutti i progetti dei programmi regionali sulla Regione di riferimento.
}
\note{
La modalità **debug** non + implementata.
}
