% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{get_simply_non_loc}
\alias{get_simply_non_loc}
\title{Semplifica le localizzazioni per progetti non localizzabili}
\usage{
get_simply_non_loc(df)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}
}
\value{
Il dataset con l le variabili x_REGIONE e x_MACROAREA modificate. In particolare distingue "PLURI-LOCALIZZATI" tra regioni della stessa macroarea e
"AMBITO NAZIONALE" per i progetti di ambito nazionale "puro" oppure con localiuzzazioni regionali molteplici trasversali alle macroare.
}
\description{
Sovrascrive le variabili x_REGIONE e x_MACROAREA semplificando le localizzazioni non reigonalizzabili.
}
\note{
La modalità **debug** non + implementata.
}
