% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Wrapper di "not in"}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Vettore da verificare.}

\item{y}{Vettore dei casi da ricercare nel vettore da verificare.}
}
\value{
Vettorelogico con lunghezza di x e valore TRUE per "not in".
}
\description{
Confronta due vettore per verificare casi "not in".
}
