% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{init_info}
\alias{init_info}
\title{Inizializza informazioni aggiuntive del database della programmazione}
\usage{
init_info(use_en = FALSE, use_713 = FALSE, sum_po = FALSE)
}
\arguments{
\item{sum_po}{Vuoi dati aggregati come summary per programma, con decisioni collassati?}

\item{usa_en}{Vuoi caricare le traduzioni in inglese?}

\item{usa_713}{Vuoi caricare anche il DB per il 2007-2013?}
}
\value{
Tutti i dati di supporto.
}
\description{
Inizializza informazioni aggiuntive del database della programmazione, con possibilità di collassare i dati per programma o tenere in evidenza singole decisioni e versioni dei programmi.
}
