% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{init_programmazione_info}
\alias{init_programmazione_info}
\title{Inizializza informazioni aggiuntive del database della programmazione}
\usage{
init_programmazione_info()
}
\value{
Tutti i dati di supporto.
}
\description{
Inizializza informazioni aggiuntive del database della programmazione, con possibilità di collassare i dati per programma o tenere in evidenza singole decisioni e versioni dei programmi.
}
