% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{init_psc}
\alias{init_psc}
\title{Inizializza PSC}
\usage{
init_psc(PSC = NULL, light = FALSE)
}
\arguments{
\item{PSC}{Folder di supporto per dati PSC}

\item{light}{Vuoi usare versione leggera (senza liste progetti art44, cds, ecc.)?}
}
\value{
Carica i file necessario al flusso nel global environment
}
\description{
Inizializza global environment per flusso di lavoro PSC
}
\note{
...
}
