% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{lista_cis_interventi_operazioni}
\alias{lista_cis_interventi_operazioni}
\title{Lista interventi cis con mapping a operazioni}
\usage{
lista_cis_interventi_operazioni(
  interventi,
  lista_clp = NULL,
  operazioni = NULL,
  progetti = NULL,
  export = TRUE
)
}
\arguments{
\item{interventi}{Interventi CIS da load_db_interventi(tipo = "CIS)}

\item{lista_clp}{Mapping tra interventi CIS e CLP...}

\item{operazioni}{Dataset da operazioni_light}

\item{progetti}{Dataset da progetti_light}

\item{export}{Vuoi esportare in TEMP in xlsx?}
}
\description{
Lista interventi cis con mapping a operazioni
}
