% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{load_correzioni_siepoc1420}
\alias{load_correzioni_siepoc1420}
\title{Carica un dataset "correzioni" dal database della programmazione}
\usage{
load_correzioni_siepoc1420(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Il dataset "correzioni".
}
\description{
Carica il dataset "correzioni" per SIE e POC 2014-2020 richiesto dal database della programmazione.
}
