% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{load_db}
\alias{load_db}
\title{Carica un dataset dal database della programmazione}
\usage{
load_db(ciclo, ambito, simplify_loc = FALSE, use_temi = FALSE)
}
\arguments{
\item{ciclo}{Ciclo di programmazione.}

\item{ambito}{Ambito di programmazione..}

\item{simplify_loc}{Logico. Vuoi semplificare le localizzazioni per compatibilità con lo standard dei Report CD?}

\item{use_temi}{Logico. Vuoi avere anche i temi prioritari FSC?}
}
\value{
Il dataset di programmazione per l'ambito richiesto, con pulizia delle denominazioni territoriali e
della codifica di aree tematiche e temi prioritari FSC.
}
\description{
Carica il datase richiesto dal dataset della programmazione.
}
