% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{load_db_accordi_ordinarie}
\alias{load_db_accordi_ordinarie}
\title{Carica dati interventi accordi per assegnazioni ordinarie}
\usage{
load_db_accordi_ordinarie(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Dataframe
}
\description{
Carica dati interventi accordi per assegnazioni ordinarie
}
