% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{load_db_altro_fsc}
\alias{load_db_altro_fsc}
\title{Carica dati altri interventi FSC}
\usage{
load_db_altro_fsc(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Dataframe
}
\description{
Carica dati altri interventi FSC
}
