% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{load_db_interventi}
\alias{load_db_interventi}
\title{Carica un dataset di tipo "interventi" dal database della programmazione}
\usage{
load_db_interventi(
  tipo,
  simplify_loc = FALSE,
  use_temi = FALSE,
  use_sog = FALSE,
  use_ue = FALSE,
  use_flt = FALSE,
  use_articolaz = FALSE,
  use_location = FALSE,
  use_ciclo = FALSE,
  tipo_ciclo = "CICLO_RISORSE"
)
}
\arguments{
\item{tipo}{Tipologia di dataset interventi.}

\item{simplify_loc}{Logico. Vuoi semplificare le localizzazioni per compatibilit? con lo standard dei Report CD?}

\item{use_temi}{Logico. Vuoi avere anche i temi prioritari FSC?}

\item{use_sog}{Logico. Vuoi avere anche il soggetto programmatore?}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{use_articolaz}{Logico. Oltre ai temi, vuoi importare anche le articolazioni? Utile per FESR e FSE}

\item{use_location}{Logico. Vuoi avere anche la localizzazione dei progetti per Regione e Macroarea?}

\item{tipo_ciclo}{Vuoi usare CICLO_STRATEGIA (default in x_AMBITO nel DB) o CICCLO_RISORSE in senso contabile (sovrascrive x_AMBITO da DB)?}

\item{use_eu}{Logico. Vuoi avere anche il finanziamento EU e categoria di regione? (solo per SIE)}
}
\value{
Il dataset di programmazione per l'ambito richiesto, con pulizia delle denominazioni territoriali e della codifica di aree tematiche e temi prioritari FSC.
}
\description{
Carica un dataset di tipo "interventi" dal database della programmazione.
}
