% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_db_pnrr}
\alias{load_db_pnrr}
\title{Carica dati di programmazione PNRR dal database della programmazione}
\usage{
load_db_pnrr(DB, use_macroaree = FALSE)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}

\item{use_macroaree}{Vuoi usare le macroaree territoriali? Default FALSE.}
}
\value{
Dataframe
}
\description{
Carica dati di programmazione PNRR dal database della programmazione
}
