% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_iter_regis}
\alias{load_iter_regis}
\title{Carica dati iter REGIS}
\usage{
load_iter_regis(scarico)
}
\arguments{
\item{scarico}{Data dello scarico in formato AAAAMMGG}
}
\value{
Dataframe con i dati degli iter
}
\description{
Carica i dati degli iter dal file CSV di REGIS
}
