% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_localizzazioni_regis}
\alias{load_localizzazioni_regis}
\title{Carica dati localizzazioni REGIS}
\usage{
load_localizzazioni_regis(scarico)
}
\arguments{
\item{scarico}{Data dello scarico in formato AAAAMMGG}
}
\value{
Dataframe con i dati delle localizzazioni
}
\description{
Carica i dati delle localizzazioni dal file CSV di REGIS
}
