% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{load_macroaree}
\alias{load_macroaree}
\title{Carica file "macroaree"}
\usage{
load_macroaree(bimestre, visualizzati = TRUE, DATA)
}
\arguments{
\item{bimestre}{Bimestre di riferimento.}

\item{visualizzati}{Logico. Vuoi solo i progetti visualizzati sul portale OC?}
}
\description{
Carica il file di base di classe "macroaree", proveniente da setup_macroaree() e salvato in DATA.
}
