% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{load_operazioni}
\alias{load_operazioni}
\title{Carica operazioni generato da workflow macroaree}
\usage{
load_operazioni(bimestre, visualizzati = TRUE, use_pqt = FALSE, DATA)
}
\arguments{
\item{visualizzati}{Logico. Vuoi solo i progetti visualizzati sul portale OC?}

\item{use_pqt}{Vuoi leggere progetti_light da file parquet?}

\item{DATA}{Path to DATA}

\item{perimetro}{Dataset di classe operazioni}
}
\description{
Carica operazioni generato da workflow macroaree
}
