% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{load_operazioni_1420}
\alias{load_operazioni_1420}
\title{Carica operazioni 1420}
\usage{
load_operazioni_1420()
}
\value{
Dataset operazinoi da dataiuku.
}
\description{
Carica operazioni 1420 da dataiku
}
