% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{load_operazioni_713}
\alias{load_operazioni_713}
\title{Carica operazioni 713}
\usage{
load_operazioni_713()
}
\value{
Dataset operazinoi da dataiuku.
}
\description{
Carica operazioni 713 da dataiku
}
