% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{load_operazioni_macroaree}
\alias{load_operazioni_macroaree}
\title{Carica file operazioni macroaree}
\usage{
load_operazioni_macroaree(bimestre, visualizzati = TRUE, DATA)
}
\arguments{
\item{bimestre}{Bimestre di riferimento.}

\item{visualizzati}{Logico. Vuoi solo i progetti visualizzati sul portale OC?}
}
\description{
Carica file operazioni sintetizzato dal file macroaree
}
