% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_pagamenti_regis}
\alias{load_pagamenti_regis}
\title{Carica dati pagamenti REGIS}
\usage{
load_pagamenti_regis(scarico)
}
\arguments{
\item{scarico}{Data dello scarico in formato AAAAMMGG}
}
\value{
Dataframe con i dati dei pagamenti
}
\description{
Carica i dati dei pagamenti dal file CSV di REGIS
}
