% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_procedure_regis}
\alias{load_procedure_regis}
\title{Carica dati procedure REGIS}
\usage{
load_procedure_regis(scarico)
}
\arguments{
\item{scarico}{Data dello scarico in formato AAAAMMGG}
}
\value{
Dataframe con i dati delle procedure
}
\description{
Carica i dati delle procedure dal file CSV di REGIS
}
