% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_progetti_pnrr}
\alias{load_progetti_pnrr}
\title{Carica dati di base per PNRR}
\usage{
load_progetti_pnrr(bimestre_pnrr, versione_pnrr)
}
\arguments{
\item{versione_pnrr}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}

\item{bimestre}{Bimestre di riferimento}
}
\value{
Dataframe
}
\description{
Carica dati di base per PNRR
}
