% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{load_progetti_psc_migrati}
\alias{load_progetti_psc_migrati}
\title{Carica dati di base per PSC}
\usage{
load_progetti_psc_migrati(bimestre, versione_psc)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{versione_psc}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}
}
\value{
Dataframe
}
\description{
Carica dati di base per PSC
}
