% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_progetti_regis}
\alias{load_progetti_regis}
\title{Carica dati progetti REGIS}
\usage{
load_progetti_regis(scarico)
}
\arguments{
\item{scarico}{Data dello scarico in formato AAAAMMGG}
}
\value{
Dataframe con i dati dei progetti
}
\description{
Carica i dati dei progetti dal file CSV di REGIS
}
