% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_struttura_pnrr}
\alias{load_struttura_pnrr}
\title{Carica e integra la struttura PNRR}
\usage{
load_struttura_pnrr(DB, progetti, export = TRUE)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto}

\item{progetti}{Dataframe dei progetti REGIS (output di load_progetti_regis)}

\item{export}{Vuoi esportare le misure mancanti in Excel? Default TRUE}
}
\value{
Dataframe con la struttura completa del PNRR
}
\description{
Carica i dati dal DB della programmazione e integra le misure mancanti dai dati di monitoraggio
}
