% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{load_target_regis}
\alias{load_target_regis}
\title{Carica dati target REGIS}
\usage{
load_target_regis(scarico)
}
\arguments{
\item{scarico}{Data dello scarico in formato AAAAMMGG}
}
\value{
Dataframe con i dati dei target
}
\description{
Carica i dati dei target dal file CSV di REGIS
}
