% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{load_totali_dbcoe}
\alias{load_totali_dbcoe}
\title{Carica i totali di riferimento dal database della programmazione}
\usage{
load_totali_dbcoe(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Il dataset con i totali di riferimento.
}
\description{
Carica i totali di riferimento dal database della programmazione.
}
