% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{make_classi_comuni}
\alias{make_classi_comuni}
\title{Classificazione per ambiti comunali}
\usage{
make_classi_comuni(pseudo, progetti, export = TRUE)
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{export}{Logico. Vuoi salvare?}
}
\value{
Un file "pseudo".
}
\description{
Classificazione per tipologie di soggetti
}
