% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{make_classi_soggetti}
\alias{make_classi_soggetti}
\title{Classificazione per soggetti}
\usage{
make_classi_soggetti(pseudo, livelli_classe = NULL, progetti, export = TRUE)
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{livelli_classe}{Nomi per factor di classe}

\item{export}{Logico. Vuoi salvare?}
}
\value{
Un file "pseudo".
}
\description{
Classificazione per tipologie di soggetti
}
