% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_progetti_fsc.R
\name{make_lista_patti}
\alias{make_lista_patti}
\title{Lista di progetti dei Patti}
\usage{
make_lista_patti(patti_list = NULL, progetti, perimetro = NULL,
  min_cp = 0, export = TRUE, bimestre = bimestre, focus = focus,
  DATA = DATA, TEMP = TEMP)
}
\arguments{
\item{patti_list}{Elenco dei Patti da esportare, in formato c("2016PATTIABR").}

\item{progetti}{File in formato "progetti" OCTK.}

\item{perimetro}{Perimetro di riferimento per ulteriore approfondimento (es. regione). Opzionale.}

\item{min_cp}{Dimensione minima dei progetti da considerare (in euro).}

\item{export}{Vuoi salvare in csv in TEMP?}

\item{bimestre}{Quale bimestre di riferimento?}

\item{focus}{Suffisso per file in export}
}
\value{
L'intero datbase dei programmazione, suddiviso in 'po_fesr', 'po_fse', 'po_fsc' e 'po_poc'.
}
\description{
Esporta la lista di progetti di uno o più Patto, in formato Report Ministro.
}
