% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{make_opendata_decisioni}
\alias{make_opendata_decisioni}
\title{Opendata per decisioni}
\usage{
make_opendata_decisioni(
  programmi = NULL,
  progetti = NULL,
  export = TRUE,
  export_xls = TRUE
)
}
\arguments{
\item{programmi}{Dati di base da workflow_programmazione().}

\item{progetti}{Dataset di tipo 'progetti' (serve per denominazioni programmi da sito e non da DB)}

\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx per ciclo e ambito in OUTPUT?}
}
\value{
File opendata con le decisioni per ambito e per i cicli 2007-2013 e 2014-2020.
}
\description{
Crea il file opendata con le decisioni dei programmi da pubblicare sul sito di OC
}
\note{
...
}
