% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_opendata_decisioni_popsc}
\alias{make_opendata_decisioni_popsc}
\title{Decisioni programmi in PSC}
\usage{
make_opendata_decisioni_popsc(export = TRUE, export_xls = TRUE)
}
\arguments{
\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx per ciclo e ambito in OUTPUT?}
}
\value{
File opendata con le decisioni per ambito e per i cicli 2007-2013 e 2014-2020.
}
\description{
Crea il file come quelli in opendata con le decisioni dei programmi originari trasformati in PSC
}
\note{
...
}
