% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimetro.R
\name{make_perimetro}
\alias{make_perimetro}
\title{Definizione del perimetro}
\usage{
make_perimetro(pseudo, export = TRUE, debug = FALSE, progetti = NULL,
  var_ls = NULL)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_CUP.
}
\description{
Ricerca progetti per settore, sotto-settore e categoria CUP a partire da input in "categorie_cup.csv".
}
