% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{make_pseudo_edit}
\alias{make_pseudo_edit}
\title{Wrapper per query editabile}
\usage{
make_pseudo_edit(progetti, query_ls = c("query_cup"), export = TRUE)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_[1], QUERY_[2], QUERY_[N] e TIPO_QUERY.
}
\description{
Wrapper per esecuzione di una lista di query a composizione libera.
}
\section{Warning}{

I valori NA sono convertiti in 0.
Usare 1 per casi certi, 2 per casi dubbi (è scartato solo 2 unico) e 9 per casi da eliminare (ad es. per PATT)
}

