% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{make_pseudo_std}
\alias{make_pseudo_std}
\title{Wrapper per query standard su CUP, UE e PO}
\usage{
make_pseudo_std(progetti, export = TRUE)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_CUP, QUERY_PO, QUERY_UE e TIPO_QUERY.
}
\description{
Wrapper per query standard su CUP, UE e PO.
}
\section{Warning}{

I dati valori NA sono convertiti in 0.
}

