% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{make_report_bimestre_coesione}
\alias{make_report_bimestre_coesione}
\title{Crea report bimestre in modalita' "coesione"}
\usage{
make_report_bimestre_coesione(programmi = NULL, export = TRUE)
}
\arguments{
\item{programmi}{Dataset in formato "programmi" da make_report_programmi_coesione()}
}
\value{
Il report bimestre.
}
\description{
Crea report sintetico bimestrale con risorse coesione calcolate su operazioni.
}
