% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{make_report_cis_sintesi}
\alias{make_report_cis_sintesi}
\title{Report CIS}
\usage{
make_report_cis_sintesi(bimestre, analisi)
}
\arguments{
\item{bimestre}{Bimestre di OpenCoesione}

\item{analisi}{Analisi interventi CIS con programmazione, attuazione e ritardi da workflow_cis()}
}
\value{
Per ogni CIS, un dossier excel in OUTPUT/dossier
}
\description{
Report per ogni CIS con diversi livelli di aggregazione
}
