% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_report_fonti_impieghi}
\alias{make_report_fonti_impieghi}
\title{Report Fonti-Impieghi}
\usage{
make_report_fonti_impieghi(use_meuro = TRUE, export = TRUE)
}
\arguments{
\item{use_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{export}{vuoi salvare il file?}
}
\value{
Il Report Fonti-Impieghi
}
\description{
Crea Report Fonti-Impieghi con sintesi di CICLO_PROGRAMMAZIONE e CICLO_RISORSE per ambito di provenienza e utilizzo.
}
