% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimestre.R
\name{make_report_programmi}
\alias{make_report_programmi}
\title{Esporta report per Programmi}
\usage{
make_report_programmi(perimetro, usa_meuro = FALSE, add_713 = FALSE,
  focus = "elab", export = FALSE)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{add_713}{Vuoi caricare anche i dati di programmaizone per il 2007-2013?}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per programma e fase procedurale.
}
\description{
Report con apertura per programma e fase procedurale rispetto al focus selezionato
}
