% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataiku.R
\name{make_report_programmi_macroaree_coesione_dataiku}
\alias{make_report_programmi_macroaree_coesione_dataiku}
\title{Esporta report per Programmi e Macroaree con dati coesione}
\usage{
make_report_programmi_macroaree_coesione_dataiku(
  perimetro,
  usa_meuro = FALSE,
  use_eu = FALSE,
  use_flt = FALSE,
  use_cicli_psc = FALSE,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE,
  show_cp = FALSE,
  export = FALSE,
  export_xls = FALSE,
  progetti = NULL,
  DB
)
}
\arguments{
\item{perimetro}{Dataset di classe macroaree}

\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro. Attenzione: per usare Meuro il perimetro deve essere in euro, viene arrotondato dopo}

\item{use_eu}{Vuoi caricare il dataset SIE del DB con le risorse UE e la categoria di regione? (solo per SIE)}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{use_cicli_psc}{Vuoi usare i dati di programmazione per cicli dei PSC?}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni?}

\item{stime_fix_siepoc}{Per correggere i dati SIE e POC 1420 con le anticipazioni vuoi usare anche le stime?}

\item{show_cp}{Logico. Vuoi calcolare anche il costo pubblico (CP)?}

\item{export}{vuoi salvare il file?}

\item{export_xls}{Vuoi salvare i file xlsx in OUTPUT?}

\item{progetti}{dataset di tipo "progetti" da utilizzare per show_cp == TRUE}
}
\value{
Un file csv con apertura per programma e fase procedurale.
}
\description{
Report con apertura per programma e fase procedurale rispetto al focus selezionato.
E' costruito su operaizoni e dati coesione.
}
\note{
Nel report restano per definizione righe con risorse 0, che derivano da errate imputazioni di macroaree.
}
