% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{make_report_risorse}
\alias{make_report_risorse}
\title{Esporta report con risorse coesione per ciclo, ambito e macroarea}
\usage{
make_report_risorse(
  use_meuro = FALSE,
  use_flt = FALSE,
  use_cicli_psc = FALSE,
  use_fix_siepoc = TRUE,
  stime_fix_siepoc = FALSE,
  force_yei = FALSE,
  export = FALSE,
  export_name = NULL
)
}
\arguments{
\item{use_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{use_cicli_psc}{Vuoi usare i dati di programmazione per cicli dei PSC?}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni?}

\item{stime_fix_siepoc}{Per correggere i dati SIE e POC 1420 con le anticipazioni vuoi usare anche le stime?}

\item{force_yei}{Logico. Vuoi forzare FSE in YEI?}

\item{export}{Vuoi salvare il file?}

\item{export_name}{Con quale filename vuoi salvare il file?}
}
\value{
Un file csv con apertura per ciclo e macroarea.
}
\description{
Esporta report con risorse coesione per ciclo, ambito e macroarea.
}
