% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_report_risorse_coesione}
\alias{make_report_risorse_coesione}
\title{Esporta report con risorse coesione per ciclo, ambito e macroarea}
\usage{
make_report_risorse_coesione(
  ciclo = NULL,
  use_meuro = FALSE,
  tipo_ciclo = "STRATEGIA",
  export = FALSE
)
}
\arguments{
\item{ciclo}{Dataset di classe perimetro.}

\item{use_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{tipo_ciclo}{Vuoi us, use}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per ciclo e macroarea.
}
\description{
Esporta report con risorse coesione per ciclo, ambito e macroarea.
}
