% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_risorse}
\alias{make_risorse}
\title{Quadro risorse programmate}
\usage{
make_risorse(usa_713 = FALSE)
}
\arguments{
\item{usa_713}{Vuoi anche le risorse 2007-2013?}
}
\value{
Tabella di sintesi per ciclo, ambito e macroarea. Per il 2007-2013, le risorse vengono inserite in forma semplificata.
}
\description{
Crea una tabella di sintesi con le risorse programmate.
}
