% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_asset.R
\docType{data}
\name{matrix_op}
\alias{matrix_op}
\title{Obiettivi di Policy 2021-2027}
\format{
A data frame
}
\source{
Team OC su regolamenti in corso di negoziazione
}
\usage{
matrix_op
}
\description{
...
}
\keyword{datasets}
