% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_asset.R
\docType{data}
\name{matrix_ra_opos}
\alias{matrix_ra_opos}
\title{Raccordo tra RA 2014-2020 e OPOS 2021-2027}
\format{
A data frame
}
\source{
Team OC su regolamenti in corso di negoziazione
}
\usage{
matrix_ra_opos
}
\description{
... RA vs OPOS
}
\details{
Le variabili sono:
\itemize{
  \item x_AMBITO: ... (CHK: solo per 14020!)
  \item \strong{COD_RA}: codice risultato atteso (TODO: da riformulare)
  \item RA_DESC: ...
  \item OP: ...
  \item \strong{OPOS}: codice obiettivo specifico 2021-2027 (TODO: da riformulare)
  \item NOTE: ...
}
}
\keyword{datasets}
