% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{oc_init_data}
\alias{oc_init_data}
\title{Copia i dati di OC da GoogleDrive}
\usage{
oc_init_data(bimestre, data_path = NULL)
}
\arguments{
\item{bimestre}{Versione dei dati di attuazione da utilizzare. Stringa in formato "20180630" come da standard per le date in OC.}

\item{data_path}{Percorso allla fonte dati (senza folder del bimestre).}
}
\value{
I file di "progetti_preesteso" sono copiati nel folder DATA.
}
\description{
Copia i dati di OC da GoogleDrive.
}
