% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{prep_dati_psc_bimestre}
\alias{prep_dati_psc_bimestre}
\title{Dati FSC per elaborazioni PSC}
\usage{
prep_dati_psc_bimestre(
  bimestre,
  versione,
  matrix_po_psc,
  po_naz,
  art44,
  matrix_1420,
  matrix_713,
  matrix_temi_settori,
  progetti = NULL,
  progetti_pub = NULL,
  versione_sgp,
  chk_today,
  forza_flag = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di riferimento, nella versione specifica per i dati PSC (es. 20211231.01 oppure 20211231), che viene ricondotto automaticamente al bimestre OC}

\item{versione}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}

\item{matrix_po_psc}{Matrice di riconciliazione PO - PSC}

\item{art44}{Dati di base per istruttoria art44 al 20200630}

\item{matrix_1420}{Matrice di riconciliazione tra temi prioritari FSC 2014-2020 e settori di intervento PSC}

\item{matrix_713}{Matrice di riconciliazione tra obiettivi specifici QSN 2007-2013 e settori di intervento PSC}

\item{matrix_temi_settori}{Dominio per aree tematiche e settori di intervento PSC}

\item{progetti}{Dataset progetti da load_progetti(visualizzati = FALSE, light = FALSE)}

\item{progetti_pub}{Dataset progetti esteso pubblicato}

\item{versione_sgp}{Versione di riferimento dei dati provenienti da SGP}

\item{chk_today}{Parametro da passare a get_stato_attuazione(), con formato "2021-02-28"}

\item{forza_flag}{Vuoi forzare il flag viasulizzazione a 7 per i duplicati tecnici da migrazione?}
}
\value{
File "dati_psc_BIMESTRE.csv" in TEMP
}
\description{
Crea il file con operazioni FSC compatibile con PSC
}
\note{
...
}
