% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimestre.R
\name{prep_perimetro_bimestre}
\alias{prep_perimetro_bimestre}
\title{Prepara perimetor progetti}
\usage{
prep_perimetro_bimestre(bimestre, usa_meuro = TRUE)
}
\arguments{
\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{perimetro}{Dataset di classe perimetro.}
}
\description{
Prepara perimetor progetti
}
