% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{prep_perimetro_bimestre_coesione}
\alias{prep_perimetro_bimestre_coesione}
\title{Prepara perimetro operazioni con dati coesione}
\usage{
prep_perimetro_bimestre_coesione(bimestre, usa_meuro = TRUE)
}
\arguments{
\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{perimetro}{Dataset di classe perimetro.}
}
\description{
Prepara perimetro operazioni con dati coesione
}
