% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{query_atp}
\alias{query_atp}
\title{Ricerca progetti per tema prioritario FSC}
\usage{
query_atp(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_ATP.
}
\description{
Ricerca progetti per area e tema prioritario FSC (ATP) a partire da input in "aree_temi_fsc.csv".
Al momento comprende tutte le delibere (non solo quelle FSC).
}
