% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{query_cipe}
\alias{query_cipe}
\title{Ricerca progetti per delibera CIPE}
\usage{
query_cipe(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_CIPE.
}
\description{
Ricerca progetti per delibera CIPE a partire da input in "delib_cipe.csv".
Al momento comprende tutte le delibere (non solo quelle FSC).
}
