% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{query_comuni}
\alias{query_comuni}
\title{Ricerca progetti per comune}
\usage{
query_comuni(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_COMUNI.
}
\description{
Ricerca progetti per comune di localizzazione a partire da input in "comuni".
}
