% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{query_keyword}
\alias{query_keyword}
\title{Ricerca progetti per parole chiave}
\usage{
query_keyword(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_KEY.
}
\description{
Ricerca progetti per parole chiave, a partire da specifici lemmi.
}
\note{
Da implementare la versione che aggrega i lemmi per per keyword ma prode una colonna per ogni keyword (e relativo debug)
}
