% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{query_patt}
\alias{query_patt}
\title{Ricerca progetti per procedura di attivazione}
\usage{
query_patt(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_PATT.
}
\description{
Ricerca progetti per procedura di attivazione a partire da input in "patt.csv".
}
