% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{query_ue}
\alias{query_ue}
\title{Ricerca progetti per campo di intervento UE}
\usage{
query_ue(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_UE.
}
\description{
Ricerca progetti per campo di intervento (2007-2013) o categoria di spesa UE (2014-2020).
}
