% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_ue_old}
\alias{query_ue_old}
\title{Ricerca progetti per categoria UE}
\usage{
query_ue_old(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_UE.
}
\description{
Ricerca progetti per campo d'intervento UE 2014-2020 e tema prioritario UE 2007-2013 a partire da input in "categorie_ue.csv".
}
\section{Warning}{

Le variabili di riferimento non sono presenti in progetti e vanno ricercate in "clp_tema_campointervento.csv".
}

