% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{refactor_ciclo}
\alias{refactor_ciclo}
\title{Refactor di x_CICLO}
\usage{
refactor_ciclo(df)
}
\arguments{
\item{df}{Dataset progetti in formato standard.}
}
\value{
Il dataset integrato.
}
\description{
Integra un perimetro di progetti con factor di x_CICLO.
}
