% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tools.R
\name{refactor_macroarea}
\alias{refactor_macroarea}
\title{Refactor di x_MACROAREA}
\usage{
refactor_macroarea(df)
}
\arguments{
\item{df}{Dataset progetti in formato standard.}
}
\value{
Il dataset integrato.
}
\description{
Integra un perimetro di progetti con factor di x_MACROAREA
}
