% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_progetti.R
\name{refactor_progetti}
\alias{refactor_progetti}
\title{Refactor per perimetro di progetti}
\usage{
refactor_progetti(perimetro)
}
\arguments{
\item{perimetro}{Dataset in formato standard.}
}
\value{
Il dataset integrato.
}
\description{
Integra un perimetro di progetti appena caricato con i factor per x_MACROAREA, x_CICLO, x_AMBITO e OC_STATO_PROCEDURALE.
}
