% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cd.R
\name{report_cd_patti}
\alias{report_cd_patti}
\title{Esporta report Patti in formato CD}
\usage{
report_cd_patti(perimetro, focus = "elab", export = FALSE)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per patto e fase procedurale.
}
\description{
Report con apertura per patto e fase procedurale rispetto al focus selezionato
}
