% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cd.R
\name{report_cd_pianinaz}
\alias{report_cd_pianinaz}
\title{Esporta report Piani nazionali FSC in formato CD}
\usage{
report_cd_pianinaz(perimetro, focus = "elab", export = FALSE)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per piano e fase procedurale.
}
\description{
Report con apertura per piano nazionale FSC e fase procedurale rispetto al focus selezionato
}
