% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cd.R
\name{report_cd_programmi}
\alias{report_cd_programmi}
\title{Esporta report per Programmi in formato CD}
\usage{
report_cd_programmi(perimetro, focus = "elab", usa_meuro = FALSE,
  export = FALSE)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per programma e fase procedurale.
}
\description{
Report con apertura per programma e fase procedurale rispetto al focus selezionato
}
