% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{report_cis_interventi}
\alias{report_cis_interventi}
\title{Report con dettaglio per cis, titolare e intervento (in risorse coesione)}
\usage{
report_cis_interventi(
  interventi,
  lista_clp = NULL,
  operazioni = NULL,
  export = TRUE
)
}
\arguments{
\item{interventi}{Interventi CIS da load_db_interventi(tipo = "CIS)}

\item{lista_clp}{Mapping tra interventi CIS e CLP...}

\item{operazioni}{Dataset da operazioni_light}

\item{export}{Vuoi esportare in TEMP in xlsx?}
}
\description{
Report con dettaglio per cis, titolare e intervento (in risorse coesione)
}
