% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_regioni.R
\name{report_regione}
\alias{report_regione}
\title{Crea report regionale}
\usage{
report_regione(regione, debug = TRUE)
}
\arguments{
\item{regione}{Regione di interesse.}

\item{debug}{Vuoi esportare i dataset intermedi di programmazione e attuazione?}
}
\value{
Il dataset con il report regionale.
}
\description{
Crea report per regione con tutti i progetti dei programmi regionali e i progetti dei programmi nazionali \
localizzati nel territorio della regione.
}
