% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_excel.R
\name{report_summarise_coe_subtot}
\alias{report_summarise_coe_subtot}
\title{COE: riepilogo con N subtotali + totale (ordine intercalato)}
\usage{
report_summarise_coe_subtot(
  df,
  key_col,
  base_levels,
  subtotals = list(),
  total_label = "Totale"
)
}
\arguments{
\item{df}{tibble/data.frame}

\item{key_col}{character(1): colonna chiave (es. "AMM_TIT")}

\item{base_levels}{character(): livelli/ordine base della chiave}

\item{subtotals}{named list: list("Etichetta1" = c(val1, val2, ...), "Etichetta2" = c(...))}

\item{total_label}{character(1): etichetta del totale (default "Totale")}
}
\value{
tibble con: key, N, RISORSE_COE, COE, p_COE, COE_IMP, p_COE_IMP, COE_PAG, p_COE_PAG
}
\description{
COE: riepilogo con N subtotali + totale (ordine intercalato)
}
\examples{
# livelli base
lst_regioni <- c("ABRUZZO","BASILICATA","CALABRIA","CAMPANIA","MOLISE",
                 "PUGLIA","SARDEGNA","SICILIA",
                 "EMILIAROMAGNA","FRIULIVG","LAZIO","LIGURIA",
                 "LOMBARDIA","MARCHE","PABOLZANO","PATRENTO",
                 "PIEMONTE","TOSCANA","UMBRIA","VALLEDAOSTA","VENETO")

# definizione subtotali
subtotals <- list(
  "Totale Sud" = c("ABRUZZO","BASILICATA","CALABRIA","CAMPANIA","MOLISE",
                   "PUGLIA","SARDEGNA","SICILIA"),
  "Totale CN"  = c("EMILIAROMAGNA","FRIULIVG","LAZIO","LIGURIA","LOMBARDIA",
                   "MARCHE","PABOLZANO","PATRENTO","PIEMONTE","TOSCANA",
                   "UMBRIA","VALLEDAOSTA","VENETO")
)

# chiamata funzione
out <- report_summarise_cp_subtot(
  df = analisi,
  key_col = "AMM_TIT",
  base_levels = lst_regioni,
  subtotals = subtotals,
  total_label = "Totale"
)

}
