% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{set_developer}
\alias{set_developer}
\title{Gestione utenti OCTK su GoogleDrive}
\usage{
set_developer(user)
}
\arguments{
\item{user}{Nome utente da cui deriva drive_root}
}
\value{
Puntamento allla root di GoogleDrive per ogni utente (drive_root)
}
\description{
Definisce il puntamento a GoogleDrive per ogni utente.
}
