% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{setup_classi_soggetti}
\alias{setup_classi_soggetti}
\title{Crea nuova base per riclassificazione per forma giuridica soggetti}
\usage{
setup_classi_soggetti(file_name = "classi_soggetti_NEW.csv")
}
\arguments{
\item{file_name}{Nome file da salvare in INPUT.}
}
\value{
Un file "classi_soggetti.csv".
}
\description{
Crea nuova base per riclassificazione per forma giuridica soggetti
}
