% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{setup_dati_pnrr}
\alias{setup_dati_pnrr}
\title{Setup dati di attuazione PNRR}
\usage{
setup_dati_pnrr(struttura, progetti, localizzazioni, iter, pagamenti)
}
\arguments{
\item{struttura}{Dataframe della struttura programmatica del PNRR (output di load_struttura_pnrr())}

\item{progetti}{Dataframe dei progetti da REGIS (output di load_progetti_regis())}

\item{localizzazioni}{Dataframe delle localizzazioni dei progetti da REGIS (output di load_localizzazioni_regis())}

\item{iter}{Dataframe dell'iter procedurale dei progetti da REGIS (output di load_iter_regis())}

\item{pagamenti}{Dataframe dei pagamenti dei progetti da REGIS (output di load_pagamenti_regis())}
}
\value{
Dataframe con dati di attuazione PNRR
}
\description{
Setup dei dati di attuazione PNRR nello standard OpenCoesione
}
