% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnrr.R
\name{setup_db_pnrr}
\alias{setup_db_pnrr}
\title{Carica dati di programmazione PNRR dal database della programmazione}
\usage{
setup_db_pnrr(DB, scarico_italiadomani, export = TRUE)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}

\item{scarico_italiadomani}{Data della versione del dataset programmazione su italiadomani in formato AAAAMMGG (https://www.italiadomani.gov.it/content/sogei-ng/it/it/catalogo-open-data/quadro-finanziario-pnrr.html).}

\item{export}{Vuoi esportare nel DBCOE?}
}
\value{
Dataframe
}
\description{
Carica dati di programmazione PNRR dal database della programmazione
}
