% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{setup_dbcoe_dati_fsc_popsc}
\alias{setup_dbcoe_dati_fsc_popsc}
\title{Crea file matrice PO-PSC per DBCOE}
\usage{
setup_dbcoe_dati_fsc_popsc(
  file_evo,
  psc_cm = FALSE,
  export = FALSE,
  export_xls = FALSE
)
}
\arguments{
\item{file_evo}{Nome del file "assegnazioni_evo_XXXXXXXX.XX.xlsx"}

\item{psc_cm}{Logico. Vuoi convertire in PSC anche i Patti delle Città metropolitane?}

\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx nel DBCOE?}
}
\value{
File con matrice PO-PSC per DBCOE per i cicli 2007-2013 e 201-2020. I dati per programma originario hanno valori al netto delle riprogrammazioni ex art. 44. Le risorse oggetto di riprogrammazione e le risorse aggiuntive parte della strategia anti COVID sono temporaneamente riportate in programmi fittizi
}
\description{
Crea file con matrice PO-PSC per DBCOE, contenente i programmi originari confluiti nei PSC
}
\note{
Salva nella versione del DBCOE risultante come DB da __oc_init__.
}
