% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{setup_dbcoe_info_fsc_popsc}
\alias{setup_dbcoe_info_fsc_popsc}
\title{Crea file info FSC per DBCOE}
\usage{
setup_dbcoe_info_fsc_popsc(file_evo, export = FALSE, export_xls = FALSE)
}
\arguments{
\item{file_evo}{Nome del file "assegnazioni_evo_XXXXXXXX.XX.xlsx"}

\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx nel DBCOE?}
}
\value{
File file info FSC per DBCOE per i cicli 2007-2013 e 2014-2020. I PSC sono interamente nel file 2014-2020. Le risorse del ciclo 2000-2006 sono interamente riassorbite nei PSC, per questo non c'è un file dedicato.
}
\description{
Crea file info FSC per DBCOE per i cicli 2000-2006, 2007-2013 e 2014-2020
Crea file info di tipo PO-PSC per DBCOE, contenente tutte le delibere dei programmi originari confluiti nei PSC
}
\note{
Salva nella versione del DBCOE risultante come DB da __oc_init__. Diversamente dalle altre funzioni dello stesso gruppo, non provede il parametro **psc_cm** perchè l'elaborazione punta direttamente si adegua direttamente al DBCOE
}
