% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{setup_dbcoe_info_fsc_psc}
\alias{setup_dbcoe_info_fsc_psc}
\title{Crea file info FSC per DBCOE}
\usage{
setup_dbcoe_info_fsc_psc(
  file_evo,
  file_temi,
  psc_cm = FALSE,
  export = FALSE,
  export_xls = FALSE
)
}
\arguments{
\item{file_evo}{Nome del file "assegnazioni_evo_XXXXXXXX.XX.xlsx"}

\item{file_temi}{Nome del file "PSC_articolazione_tematiche_chk_decimali_v.XX.xlsx"}

\item{psc_cm}{Logico. Vuoi convertire in PSC anche i Patti delle Città metropolitane?}

\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx nel DBCOE?}
}
\value{
File file info FSC per DBCOE per i cicli 2007-2013 e 2014-2020. I PSC sono interamente nel file 2014-2020. Le risorse del ciclo 2000-2006 sono interamente riassorbite nei PSC, per questo non c'è un file dedicato.
}
\description{
Crea file info FSC per DBCOE per i cicli 2000-2006, 2007-2013 e 2014-2020
}
\note{
Salva nella versione del DBCOE risultante come DB da __oc_init__.
}
