% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{setup_dbcoe_no_fsc}
\alias{setup_dbcoe_no_fsc}
\title{#' Lista programmi non monitorabili
#'
#' Crea la lista dei programmi programmi non monitorabili
#' 
#' @param programmi Dati di base da workflow_programmazione().
#' @param progetti Dataset di tipo 'progetti' (serve per denominazioni programmi da sito e non da DB)
#' @param export vuoi salvare il file?
#' @return Lista dei programmi non monitorabili
make_lista_nomonit <- function(programmi=NULL, progetti=NULL, export=TRUE){
  
  
  
  
  if (is.null(programmi)) {
    if (is.null(progetti)) {
      progetti <- load_progetti(bimestre, visualizzati=TRUE, light=TRUE)
    }
    programmi <- workflow_programmazione(use_info=TRUE, use_flt=TRUE, progetti)
  }
  
  info_en <- init_programmazione_info(use_en = TRUE, use_713 = TRUE, sum_po = TRUE, sum_po_last = TRUE, use_po_psc = TRUE) %>%
    select(OC_CODICE_PROGRAMMA, LABEL_DECISIONE_EN)
  
  # programmi_en <- read_csv2(file.path(INPUT, "programmi_SIE_EN.csv")) %>%
  programmi_en <- read_xlsx(file.path(DB, "label_programmi_en.xlsx")) %>% 
    # select(-LABEL_PROGRAMMA_IT)
    distinct(OC_CODICE_PROGRAMMA, LABEL_PROGRAMMA_EN)
  
  # DEV:
  # psc <- programmi %>% 
  #   filter(x_GRUPPO == "PSC") %>% 
  #   distinct(OC_CODICE_PROGRAMMA)
  # 
  # temp <- read_xlsx(file.path(DB, "fsc_delibere_psc.xlsx")) %>% 
  #   mutate(NUMERO_DECISIONE = as.character(NUMERO_DECISIONE)) %>% 
  #   rename(AMBITO.x = AMBITO)
  # 
  # info_2 <- info %>% 
  #   anti_join(psc) %>% 
  #   bind_rows(temp)
  
  # label LABEL_PROGRAMMA_EN
  programmi <- programmi %>%
    left_join(info_en, by = "OC_CODICE_PROGRAMMA") %>%
    left_join(programmi_en, by = "OC_CODICE_PROGRAMMA") %>%
    mutate(LABEL_PROGRAMMA_IT = x_PROGRAMMA,
           LABEL_PROGRAMMA_EN = if_else(is.na(LABEL_PROGRAMMA_EN), LABEL_PROGRAMMA_IT, LABEL_PROGRAMMA_EN))
  # CHK: duplicato, programmi incrementa di 4
  
  # temp <- programmi %>% count(OC_CODICE_PROGRAMMA) %>% filter(n > 1)
  # chk <- programmi %>% semi_join(temp)
  # print("Controlla se numerosità è invariata!!!")
  # dim(programmi)[1] == dim(programmi_base)[1]
  # PON Inclusione
  
  # make URL_PROGRAMMA 
  programmi <- programmi %>%
    mutate(LINK_PROGRAMMA_IT = if_else(PUB == TRUE,
                                       paste0("https://opencoesione.gov.it/it/programmi/", OC_CODICE_PROGRAMMA, "/"),
                                       ""),
           LINK_PROGRAMMA_EN = if_else(PUB == TRUE,
                                       paste0("https://opencoesione.gov.it/en/programmi/", OC_CODICE_PROGRAMMA, "/"),
                                       "")) %>% 
    mutate(LINK_DOC_IT = paste0("https://opencoesione.gov.it/it/programmi/", OC_CODICE_PROGRAMMA, "/documenti/"),
           LINK_DOC_EN = paste0("https://opencoesione.gov.it/en/programmi/", OC_CODICE_PROGRAMMA, "/documenti/"))
  
  
  # LABEL_SITO_IT
  programmi <- programmi %>%
    mutate(LABEL_SITO_IT = if_else(is.na(LINK_SITO), "", "Sito web"),
           LABEL_SITO_EN = if_else(is.na(LINK_SITO), "", "Website"),
           LABEL_DOC_IT  = if_else(is.na(LINK_DOC_IT), "", "Documenti"),
           LABEL_DOC_EN  = if_else(is.na(LINK_DOC_EN), "", "Documents"),)
  
  programmi <- programmi %>% 
    mutate(LINK_DECISIONE = case_when(x_AMBITO == "FSC" ~ LINK_DECISIONE,
                                      x_AMBITO == "POC" ~ LINK_DECISIONE,
                                      x_AMBITO == "PAC" ~ LINK_DECISIONE,
                                      TRUE ~ ""))
  
  # elimina "da programmare" e "non coesione" (fonte DEF 2020)
  # programmi <- programmi %>%
  #   filter(OC_CODICE_PROGRAMMA != "FSC_1420_CDD_40", # MEMO: da programmare FSC
  #          OC_CODICE_PROGRAMMA != "2020DAPROGR_1", # MEMO: sicilia e anpal
  #          OC_CODICE_PROGRAMMA != "2020DAPROGR_2", 
  #          OC_CODICE_PROGRAMMA != "TEMP_0713_007", # MEMO: debiti regioni 713
  #          OC_CODICE_PROGRAMMA != "DEBITI_CAM")
  # dovrà essere superato con OC_FLAG_MONITORAGGIO in init_programmazione_dati
  
  
  # label LABEL_AMBITO_EN
  programmi <- programmi %>%
    mutate(x_AMBITO = as.character(x_AMBITO)) %>% 
    mutate(LABEL_AMBITO_IT = case_when(x_AMBITO == "YEI" ~ "IOG",  # fix YEI
                                       x_AMBITO == "SNAI" ~ "SNAI-Servizi",
                                       TRUE ~ x_AMBITO),
           LABEL_AMBITO_EN = case_when(x_AMBITO == "FESR" ~ "ERDF",
                                       x_AMBITO == "FSE" ~ "ESF",
                                       x_AMBITO == "FEASR" ~ "EAFRD",
                                       x_AMBITO == "FEAMP" ~ "MFF",
                                       x_AMBITO == "CTE" ~ "ETC",
                                       x_AMBITO == "FSC" ~ "DCF",
                                       x_AMBITO == "POC" ~ "COP",
                                       x_AMBITO == "SNAI" ~ "IANS",
                                       x_AMBITO == "YEI" ~ "YEI"),
           LABEL_TIPO_IT = x_GRUPPO,
           LABEL_TIPO_EN = case_when(x_GRUPPO == "PON" ~ "NOP",
                                     x_GRUPPO == "POR" ~ "ROP",
                                     x_GRUPPO == "PATTI" ~ "DEVELOPMENT PACT",
                                     x_GRUPPO == "PIANI STRALCIO" ~ "EXCERPT PLAN",
                                     x_GRUPPO == "PIANI OPERATIVI" ~ "NATIONAL PLAN",
                                     x_GRUPPO == "POC REGIONALI" ~ "REGIONAL COP",
                                     x_GRUPPO == "POC NAZIONALI" ~ "NATIONAL COP",
                                     x_GRUPPO == "SNAI" ~ "IANS",
                                     TRUE ~ x_GRUPPO))
  
  # maiusc
  programmi <- programmi %>% 
    mutate(LABEL_PROGRAMMA_IT = toupper(LABEL_PROGRAMMA_IT),
           LABEL_PROGRAMMA_EN = toupper(LABEL_PROGRAMMA_EN))
  
  # FIX nomi psc ministeri creativi
  # programmi <- programmi %>% 
  #   mutate(LABEL_PROGRAMMA_IT = case_when(OC_CODICE_PROGRAMMA == "PSC_MIT" ~ "PSC MINISTERO INFRASTRUTTURE E MOBILITA' SOSTENIBILE",
  #                                         OC_CODICE_PROGRAMMA == "PSC_MATTM" ~ "PSC MINISTERO TRANSIZIONE ECOLOGICA",
  #                                         TRUE ~ LABEL_PROGRAMMA_IT))
  
  # export #
  out <- programmi %>%
    filter(LABEL_AMBITO_IT != "FEASR", LABEL_AMBITO_IT != "FEAMP") %>% 
    mutate(LABEL_TIPO_IT = case_when(LABEL_AMBITO_IT == "FSC" & LABEL_TIPO_IT == "PATTI" ~ "PATTI",
                                     LABEL_AMBITO_IT == "FSC" & LABEL_TIPO_IT == "PSC" ~ "PSC",
                                     LABEL_AMBITO_IT == "FSC" ~ "VARI",
                                     LABEL_AMBITO_IT == "POC" & LABEL_TIPO_IT == "POC Nazionale" ~ "NAZIONALI",
                                     LABEL_AMBITO_IT == "POC" & LABEL_TIPO_IT == "POC Nazionale Completamenti" ~ "COMPLETAMENTI",
                                     LABEL_AMBITO_IT == "POC" & LABEL_TIPO_IT == "POC Regionale" ~ "REGIONALI",
                                     LABEL_AMBITO_IT == "POC" & LABEL_TIPO_IT == "POC Regionale Completamenti" ~ "COMPLETAMENTI",
                                     LABEL_AMBITO_IT == "SNAI-Servizi" ~ "SNAI-SERVIZI",
                                     TRUE ~ LABEL_TIPO_IT)) %>% 
    mutate(LABEL_TIPO_IT = factor(LABEL_TIPO_IT, levels = c("PSC", "PATTI", "VARI", "POR", "PON", "CTE", "NAZIONALI", "REGIONALI", "COMPLETAMENTI", "SNAI-SERVIZI"))) %>%
    mutate(LABEL_TIPO_EN = case_when(LABEL_AMBITO_EN == "DCF" & LABEL_TIPO_EN == "PACTS" ~ "PACTS",
                                     LABEL_AMBITO_EN == "DCF" & LABEL_TIPO_EN == "PSC" ~ "PSC",
                                     LABEL_AMBITO_EN == "DCF" ~ "OTHERS",
                                     LABEL_AMBITO_IT == "COP" & LABEL_TIPO_IT == "POC Nazionale" ~ "NAZIONALI",
                                     LABEL_AMBITO_IT == "COP" & LABEL_TIPO_IT == "POC Nazionale Completamenti" ~ "COMPLETAMENTI",
                                     LABEL_AMBITO_IT == "COP" & LABEL_TIPO_IT == "POC Regionale" ~ "REGIONALI",
                                     LABEL_AMBITO_IT == "COP" & LABEL_TIPO_IT == "POC Regionale Completamenti" ~ "COMPLETAMENTI",
                                     TRUE ~ LABEL_TIPO_EN)) %>% 
    mutate(LABEL_TIPO_EN = factor(LABEL_TIPO_EN, levels = c("PSC", "PACTS", "OTHERS", "NOR", "NOP", "CTE", "NAZIONALI", "REGIONALI", "COMPLETAMENTI", "SNAI-SERVIZI"))) %>%
    mutate(LINK_DOC = paste0("../programmi/", OC_CODICE_PROGRAMMA, "/documenti/")) %>% # TEST
    mutate(RISORSE = round(RISORSE, 0),
           RISORSE_UE = round(RISORSE_UE, 0)) %>% 
    mutate(LINK_DOC = case_when(OC_CODICE_PROGRAMMA == "TEMP_CTE_TRANS	" ~ "", #sere per non generare link su sito
                                OC_CODICE_PROGRAMMA == "COMP_POC_CALABR" ~ "",
                                OC_CODICE_PROGRAMMA == "COMP_POC_CAMPAN" ~ "",
                                OC_CODICE_PROGRAMMA == "COMP_POC_CULTUR" ~ "",
                                OC_CODICE_PROGRAMMA == "COMP_POC_ENERGI" ~ "",
                                OC_CODICE_PROGRAMMA == "COMP_POC_SICILI" ~ "",
                                OC_CODICE_PROGRAMMA == "COMP_POC_LEGALI" ~ "",
                                OC_CODICE_PROGRAMMA == "AREEINTASSTEC" ~ "",
                                TRUE ~ LINK_DOC)) %>% 
    mutate(LABEL_DOC_EN = case_when(OC_CODICE_PROGRAMMA == "TEMP_CTE_TRANS	" ~ "", #sere per non generare link su sito
                                    OC_CODICE_PROGRAMMA == "COMP_POC_CALABR" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_CAMPAN" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_CULTUR" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_ENERGI" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_SICILI" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_LEGALI" ~ "",
                                    OC_CODICE_PROGRAMMA == "AREEINTASSTEC" ~ "",
                                    TRUE ~ LABEL_DOC_EN)) %>% 
    mutate(LABEL_DOC_IT = case_when(OC_CODICE_PROGRAMMA == "TEMP_CTE_TRANS	" ~ "", #sere per non generare link su sito
                                    OC_CODICE_PROGRAMMA == "COMP_POC_CALABR" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_CAMPAN" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_CULTUR" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_ENERGI" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_SICILI" ~ "",
                                    OC_CODICE_PROGRAMMA == "COMP_POC_LEGALI" ~ "",
                                    OC_CODICE_PROGRAMMA == "AREEINTASSTEC" ~ "",
                                    TRUE ~ LABEL_DOC_IT)) %>% 
    # deve stare dopo
    mutate(OC_CODICE_PROGRAMMA = case_when(OC_CODICE_PROGRAMMA == "TEMP_CTE_TRANS	" ~ "", #sere per non generare link su sito
                                           OC_CODICE_PROGRAMMA == "COMP_POC_CALABR" ~ "",
                                           OC_CODICE_PROGRAMMA == "COMP_POC_CAMPAN" ~ "",
                                           OC_CODICE_PROGRAMMA == "COMP_POC_CULTUR" ~ "",
                                           OC_CODICE_PROGRAMMA == "COMP_POC_ENERGI" ~ "",
                                           OC_CODICE_PROGRAMMA == "COMP_POC_SICILI" ~ "",
                                           OC_CODICE_PROGRAMMA == "COMP_POC_LEGALI" ~ "",
                                           OC_CODICE_PROGRAMMA == "AREEINTASSTEC" ~ "",
                                           TRUE ~ OC_CODICE_PROGRAMMA)) %>% 
    select(OC_CODICE_PROGRAMMA,
           LABEL_PROGRAMMA_IT,
           LABEL_PROGRAMMA_EN,
           LABEL_CICLO = x_CICLO,
           LABEL_AMBITO_IT,
           LABEL_AMBITO_EN,
           LABEL_TIPO_IT,
           LABEL_TIPO_EN,
           RISORSE,
           RISORSE_UE,
           LABEL_DECISIONE_IT,
           LABEL_DECISIONE_EN,
           LINK_DECISIONE,
           LABEL_DOC_IT,
           LABEL_DOC_EN,
           LINK_DOC, # TEST
           # LINK_DOC_IT, 
           # LINK_DOC_EN,
           LABEL_SITO_IT,
           LABEL_SITO_EN,
           # LINK_PROGRAMMA_IT, #generate automaticamente su sito
           # LINK_PROGRAMMA_EN, #generate automaticamente su sito
           LINK_SITO
           
    ) %>% 
    arrange(LABEL_TIPO_IT)
  
  # split cicli
  out_1420 <- out %>% filter(LABEL_CICLO == "2014-2020")
  out_713 <- out %>% filter(LABEL_CICLO == "2007-2013")
  
  if (export == TRUE) {
    require(withr)
    withr::with_options(
      c(scipen = 10), 
      write.csv2(out_1420, file.path(TEMP, "programmi_1420.csv"), row.names = FALSE, na = "")
    )
    withr::with_options(
      c(scipen = 10), 
      write.csv2(out_713, file.path(TEMP, "programmi_0713.csv"), row.names = FALSE, na = "")
    )
  }
  
  return(out)
}
Copia file DBCOE in nuova versione}
\usage{
setup_dbcoe_no_fsc(db_path_old, db_path_new = NULL)
}
\arguments{
\item{db_path_old}{Path alla versione più antica del DBCOE.}

\item{db_path_new}{Path alla versione più recente del DBCOE.}
}
\value{
File DBCOE nel folder della nuova versione
}
\description{
Copia file DBCOE dalla cartella di una versione precedente a quella di nuova versione, per avviare il setup. Copia tutto tranne FSC.
}
