% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{setup_light}
\alias{setup_light}
\title{Crea progetti_light.csv}
\usage{
setup_light(bimestre, fix = FALSE, ...)
}
\arguments{
\item{bimestre}{Bimestre di riferimento.}
}
\value{
Il dataset viene salvato in DATA e può essere caricato con load_progetti(light = TRUE).
}
\description{
Crea un file progetti da progetti_esteso.csv, con le sole variabili fondamentali. Aggiunge blocco x_var e x_MACROAREA.
}
