% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{setup_macroaree}
\alias{setup_macroaree}
\title{Crea il dataset macroaree}
\usage{
setup_macroaree(
  bimestre,
  progetti,
  operazioni_713,
  operazioni_1420,
  export = FALSE,
  use_fix = FALSE,
  debug = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{progetti}{Dataset con un perimetro in formato "progetti".}

\item{operazioni_713}{File di tipo operazioni da flusso sas/dataiku.}

\item{operazioni_1420}{File di tipo operazioni da flusso sas/dataiku.}

\item{export}{Vuoi esportare in DATA?}

\item{use_fix}{Vuoi applicare fix_progetti() prima di esecuzione? Non impatta su x_AMBITO!}
}
\value{
Il dataset macroaree con le variabili coesione calcolate per articolazione e macroarea territoriale.
}
\description{
Crea il dataset macroaree con operazioni per articolazioni e macroaree territoriali.
}
\note{
La modalità **debug** esporta diversi csv in TEMP La modalità **export** esporta macroaree_light.csv in DATA.
}
