% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_psc.R
\name{setup_macroaree_psc_migrati}
\alias{setup_macroaree_psc_migrati}
\title{Crea file di base per macroaree}
\usage{
setup_macroaree_psc_migrati(progetti_psc, operazioni, export = FALSE)
}
\arguments{
\item{progetti_psc}{Dataset da load_progetti_psc_migrati()}

\item{operazioni}{Dataset da load_operazioni(), con visualizzati=FALSE}

\item{export}{vuoi salvare il file?}
}
\value{
Dataset macroaree per il report make_report_temi_macroaree_psc_migrati
}
\description{
Crea file di base per macroaree, da operazioni, per psc migrati.
}
