% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{setup_operazioni}
\alias{setup_operazioni}
\title{Setup di operazioni da workflow macroaree}
\usage{
setup_operazioni(
  bimestre,
  progetti,
  operazioni_713_raw,
  operazioni_1420_raw,
  operazioni_extra_raw,
  export = TRUE,
  export_pqt = FALSE,
  debug = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di rifeirimento da oc_init().}

\item{progetti}{Dataset "progetti" in formato PREESTESO per integrazione.}

\item{operazioni_713_raw}{File di tipo operazioni da flusso sas/dataiku.}

\item{operazioni_1420_raw}{File di tipo operazioni da flusso sas/dataiku.}

\item{operazioni_extra_raw}{File di tipo operazioni da flusso sas/dataiku.}

\item{export}{Vuoi esportare il file in formato operazioni?}

\item{export_pqt}{Vuoi esportare il file in formato operazioni in formato parquet?}

\item{debug_mode}{Vuoi esportare i file di debug?}
}
\value{
Il dataset operazioni.
}
\description{
Setup di operazioni da workflow macroaree
}
