% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{setup_operazioni_macroaree}
\alias{setup_operazioni_macroaree}
\title{Crea il dataset operazioni}
\usage{
setup_operazioni_macroaree(bimestre, macroaree, export = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{macroaree}{Dataset di tipo "macroaree".}

\item{export}{Vuoi esportare in DATA?}
}
\value{
Il dataset operazioni con le variabili coesione calcolate: COE, COE_IMP e COE_PAG.
}
\description{
Crea il dataset operazioni a partire dal dataset macroaree_light.
}
\note{
La modalità **export** esporta operazioni_light.csv in DATA.
}
