% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{setup_progetti_descr}
\alias{setup_progetti_descr}
\title{Crea descrizione_progetti.csv}
\usage{
setup_progetti_descr(bimestre, fix = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento.}
}
\value{
Il dataset viene salvato in DATA.
}
\description{
Crea un file con le sole descrizionini dei progetti da progetti_esteso.csv.
}
