% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{setup_progetti_light}
\alias{setup_progetti_light}
\title{Crea progetti_light.csv}
\usage{
setup_progetti_light(bimestre, progetti, operazioni_713, fix = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento.}

\item{progetti}{File di tipo progetti da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{operazioni_713}{File di tipo operazioni da flusso sas/dataiku.}
}
\value{
Il dataset viene salvato in DATA e può essere caricato con load_progetti(light = TRUE).
}
\description{
Crea un file progetti da progetti_esteso.csv, con le sole variabili fondamentali. Aggiunge blocco x_var e x_MACROAREA.
}
