% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritardi.R
\name{setup_ritardi}
\alias{setup_ritardi}
\title{Setup analisi ritardi}
\usage{
setup_ritardi(
  bimestre,
  progetti,
  progetti_old,
  chk_today,
  perimetro = NULL,
  debug = FALSE,
  export = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di OpenCoesione}

\item{progetti}{Dataset da load_progetti(bimestre, light = FALSE, visualizzati = FALSE)}

\item{progetti_old}{Dataset da progetti_light riferito a un bimestre precedente (serve per controllo variazioni date previste)}

\item{chk_today}{Parametro da passare a get_stato_attuazione(), con formato "2021-02-28"}

\item{perimetro}{Elenco di CLP per focus specifico}

\item{debug}{Vuoi verificare ...?}

\item{export}{Vuoi esportare in TEMP in xlsx?}
}
\value{
Dataset di base per analisi ritardi iter procedurale
}
\description{
Crea dataset di base per analisi ritardi iter procedurale. Trova fase corrente, verifica ritardo su data prevista per prossima data effettiva assnte.
Ricalcola x_STATO_PROCEDURALE con get_stato_attuazione()
}
