% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_excel.R
\docType{data}
\name{style_border}
\alias{style_border}
\title{Stile per celle con bordo}
\format{
An object of class \code{Style} of length 1.
}
\usage{
style_border
}
\value{
Stile compatibile con addStyle() in openxlsx
}
\description{
Stile per celle con bordo da applicare con openxlsx
}
\note{
Non prevede parametri editabili, sono stili standard per OC
}
\keyword{datasets}
