% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_excel.R
\docType{data}
\name{style_date2}
\alias{style_date2}
\title{Stile per celle con date senza bordo}
\format{
An object of class \code{Style} of length 1.
}
\usage{
style_date2
}
\value{
Stile compatibile con addStyle() in openxlsx
}
\description{
Stile per celle con date senza bordo da applicare con openxlsx
}
\note{
Non prevede parametri editabili, sono stili standard per OC
}
\keyword{datasets}
