% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_perimetri.R
\name{update_input_with_delta}
\alias{update_input_with_delta}
\title{Crea file di input integrato con delta da precedente elaborazione}
\usage{
update_input_with_delta(OLD)
}
\arguments{
\item{OLD}{Percorso alla cartella di livello WORK della precedente elaborazione presa a riferimento.}
}
\value{
Esegue bind del precedente input e della versione delta, salvando un file "input_query.xlsx".
}
\description{
Crea file di input integrato con input di precedente elaborazione e delta rispetto all'attuale template nel package, da calcolare con \link[octk]{make_input_delta}.
}
