% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{update_lista_programmi_en}
\alias{update_lista_programmi_en}
\title{Update lista denominazione programmi in inglese}
\usage{
update_lista_programmi_en(db_old, progetti = NULL)
}
\arguments{
\item{db_old}{Versione precedente del DBCOE, nel format standard tipo 20221231.01}

\item{progetti}{Dataset progetti importato con load_progetti}
}
\value{
Il file "label_programmi_en.xlsx" viene copiato dalla versione precedente a quella corrente e integrato con i nuovi programmi
}
\description{
Update lista denominazione programmi in inglese.
}
