% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{update_lista_programmi_sitiweb}
\alias{update_lista_programmi_sitiweb}
\title{Update lista siti web programmi}
\usage{
update_lista_programmi_sitiweb(db_old)
}
\arguments{
\item{db_old}{Versione precedente del DBCOE, nel format standard tipo 20221231.01}
}
\value{
Il file "label_programmi_en.xlsx" viene copiato dalla versione precedente a quella corrente e integrato con i nuovi programmi
}
\description{
Update lista siti web programmie.
}
