% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{workflow_chk_bimestre}
\alias{workflow_chk_bimestre}
\title{Workflow chk bimestre}
\usage{
workflow_chk_bimestre(progetti, export = FALSE)
}
\arguments{
\item{progetti}{File di tipo progetti per il bimestre corrente da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{export}{Vuoi esportare i report di controllo in TEMP?}
}
\value{
Messaggi e report di controllo.
}
\description{
Workflow di controllo dei dati del nuovo bimestre
}
