% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{workflow_cis}
\alias{workflow_cis}
\title{Workflow per analisi CIS}
\usage{
workflow_cis(
  bimestre,
  interventi,
  operazioni,
  ritardi = NULL,
  debug = FALSE,
  export = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di OpenCoesione}

\item{interventi}{Interventi CIS da load_db_interventi(tipo = "CIS)}

\item{operazioni}{Dataset da operazioni_light}

\item{ritardi}{Dataset da setup_ritardi()}

\item{debug}{Vuoi verificare il DB e il mappging con l'attuazione?}

\item{export}{Vuoi esportare in TEMP in xlsx?}

\item{progetti}{Dataset da progetti_light}
}
\value{
Dataset di base per analisi programmazione vs attuazione CIS
}
\description{
Crea dataset di base per analisi programmazione vs attuazione CIS
}
