% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{workflow_macroaree_sub_pivot}
\alias{workflow_macroaree_sub_pivot}
\title{Workflow per creazione file di base per macroaree}
\usage{
workflow_macroaree_sub_pivot(df)
}
\arguments{
\item{df}{Dataset di classe operazioni}
}
\value{
Dataset pivot con macroaree per setup_macroaree_sie e setup_macroaree_psc
}
\description{
Workflow per creazione file di base per macroaree.
Funziona con input specifici per sie non react e per psc migrati.
}
