% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_operazioni.R
\name{workflow_macroaree_sub_programmazione}
\alias{workflow_macroaree_sub_programmazione}
\title{Integra la macroarea da localizzazioni con correzioni per programmazione}
\usage{
workflow_macroaree_sub_programmazione(df, df_raw, progetti)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti" in formato PREESTESO per integrazione.}
}
\value{
Il dataset con la variabile x_MACROAREA, come factor con levels = c("Centro-Nord", "Sud", "Trasversale", "Nazionale", "Estero").
}
\description{
Integra in x_MACROAREA la macroarea dalle localizzazioni di PREESTESO partendo da OC_MACROAREA. 
Corregge per i programmi regionali forzando l'attribuzione alla regione di rifeirmento del programma.
}
