% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{workflow_macroaree_sub_psc}
\alias{workflow_macroaree_sub_psc}
\title{Apre operazioni psc per macroarea}
\usage{
workflow_macroaree_sub_psc(operazioni, debug = FALSE)
}
\arguments{
\item{operazioni}{Dataset di classe operazioni interno al flusso di workflow_macroaree().}

\item{debug}{vuoi salvare il file?}
}
\value{
Dataset macroaree psc per il workflow.
}
\description{
Apre operazioni per macroarea, per programmi psc, all'interno del flusso di workflow_macroaree(), in setup_macroaree().
}
