% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{workflow_macroaree_sub_sie}
\alias{workflow_macroaree_sub_sie}
\title{Apre operazioni sie per macroarea}
\usage{
workflow_macroaree_sub_sie(operazioni, riprop_yei = FALSE, debug = FALSE)
}
\arguments{
\item{operazioni}{Dataset di classe operazioni interno al flusso di workflow_macroaree().}

\item{riprop_yei}{Vuoi riproporzionare yei dove mancano dati finanziari per livelli gerarchici? Default su FALSE.}

\item{debug}{vuoi salvare il file?}
}
\value{
Dataset macroaree psc per il workflow.
}
\description{
Apre operazioni per macroarea, per progetti sie non react e non yei, all'interno del flusso di workflow_macroaree(), in setup_macroaree().
Per react e yei usa macroarea da localizzazioni.
}
