% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{workflow_operazioni}
\alias{workflow_operazioni}
\title{Workflow per creare il dataset operazioni}
\usage{
workflow_operazioni(bimestre, progetti = NULL, debug = FALSE,
  use_fix = FALSE)
}
\description{
Workflow per creare il dataset operazioni
}
