% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataiku.R
\name{workflow_operazioni_dataiku}
\alias{workflow_operazioni_dataiku}
\title{Workflow per creare il dataset operazioni}
\usage{
workflow_operazioni_dataiku(
  bimestre,
  progetti,
  operazioni_713,
  operazioni_1420,
  debug = FALSE
)
}
\description{
Workflow per creare il dataset operazioni
}
