% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{workflow_operazioni_ecomix}
\alias{workflow_operazioni_ecomix}
\title{Workflow per creare il dataset operazioni}
\usage{
workflow_operazioni_ecomix(bimestre, progetti, debug = FALSE)
}
\description{
Workflow per creare il dataset operazioni
}
