% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{workflow_pivot_macroaree}
\alias{workflow_pivot_macroaree}
\title{Workflow per creazione file di base per macroaree}
\usage{
workflow_pivot_macroaree(operazioni)
}
\arguments{
\item{perimetro}{Dataset di classe operazioni}
}
\value{
Dataset pivot con macroaree per setup_macroaree_sie e setup_macroaree_psc
}
\description{
Workflow per creazione file di base per macroaree.
Funziona con input specifici per sie non react e per psc migrati.
}
