% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_programmazione.R
\name{workflow_programmazione}
\alias{workflow_programmazione}
\title{Workflow di preparazione della programmazione per la pubblicazione.}
\usage{
workflow_programmazione(
  use_flt = TRUE,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE,
  use_location = FALSE,
  progetti
)
}
\arguments{
\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{use_fix_siepoc}{Logico. Vuoi correggere i dati SIE e POC 1420 con le anticipazioni?}

\item{stime_fix_siepoc}{Per correggere i dati SIE e POC 1420 con le anticipazioni vuoi usare anche le stime?}

\item{use_location}{Logico. Vuoi mostreare le macroaree territoriali? Serve per opendata su dotazioni.}

\item{progetti}{Dataset progetti importato con load_progetti}
}
\value{
Il dataset dei programmi con risorse e evewntualmente infomrmazioni di supporto
}
\description{
Applica convenzioni per pubblicazione su OpenCoesione, che non possono essere direttamente riportate nel DBCOE.
}
