% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{workflow_setup_rawdata}
\alias{workflow_setup_rawdata}
\title{Workflow integrazione raw data}
\usage{
workflow_setup_rawdata(RAWDATA)
}
\arguments{
\item{RAWDATA}{Percorso al repository (da file.path(DRIVE, "DATI", "OCTK", "_dat"))}
}
\value{
Messaggi e report di controllo.
}
\description{
Workflow integrazione raw data per nuovo bimestre. Rende disponibili nel package in formato .Rdata i file nel repository in TOOLS > OCTK > _dat.
}
