% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_setup.R
\name{workflow_update_rawdata}
\alias{workflow_update_rawdata}
\title{Workflow integrazione raw data}
\usage{
workflow_update_rawdata(progetti, RAWDATA)
}
\arguments{
\item{progetti}{File di tipo progetti per il bimestre corrente da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{RAWDATA}{Percorso al repository (da file.path(DRIVE, "DATI", "OCTK", "_dat"))}
}
\value{
File nel repository in TOOLS > OCTK > _dat.
}
\description{
Workflow integrazione raw data per nuovo bimestre. Aggirona i file csv nel repository in TOOLS > OCTK > _dat.
}
