% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{add_perimetro_to_pseudo}
\alias{add_perimetro_to_pseudo}
\title{Aggiunge progetti a pseudo da un perimetro}
\usage{
add_perimetro_to_pseudo(
  pseudo,
  addendum,
  usa_classi = FALSE,
  add_name = "QUERY_ADD",
  export = TRUE
)
}
\arguments{
\item{pseudo}{Dataset in formato "pseudo"}

\item{addendum}{Dataset in memory oppure nome del file in _OUTPUT_SAS da caricare in INPUT, con elenco di COD_LOCALE_PROGETTO}

\item{usa_classi}{Logico. Se il perimetro è diviso in classi vuoi tracciarle separatamente in diverse colonne di tipo QUERY_XXX?}

\item{add_name}{Nome variabile di tipo QUERY_XXX}

\item{export}{Logico. Vuoi salvare pseudo.csv?}
}
\value{
Un dataframe pseudo integrato con la colonna QUERY_ADD, che si applica anche a righe esistenti (la funzione usa full_join e non bind_rows!).
}
\description{
Integra pseudo csv con righe provenienti da selezione proveniente da un perimetro realizzato con il package, e in caso di uso di CLASSI le trasfroma in diversi QUERY_XXX.
}
\section{Warning}{

La funzione è fuori dal blocco "query" solo per maggiore trasparenza, si poteva usare anche quella logica.
}

