% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{add_to_pseudo}
\alias{add_to_pseudo}
\title{Addendum Turismo}
\usage{
add_to_pseudo(pseudo, addendum, export = TRUE)
}
\arguments{
\item{pseudo}{Dataset in formato "pseudo"}

\item{addendum}{Dataset in memory oppure nome del file da caricare in INPUT, con elenco di COD_LOCALE_PROGETTO}

\item{export}{Logic. Vuoi salvare pseudo.csv?}
}
\value{
Un dataframe pseudo.

Un dataframe pseudo integrato con la colonna QUERY_ADD, che si applica anche a righe esistenti (la funzione usa full_join e non bind_rows!).
}
\description{
....
}
\section{Warning}{

Punta a "old_perim.csv".
Forse non serve nemmeno più...
Aggiunge progetti a pseudo da una lista di CLP

Integra pseudo csv con righe provenienti da selezione proveniente da fonte esterna alle funzioni del package e crea colonna "query_add".


La funzione è fuori dal blocco "query" solo per maggiore trasparenza, si poteva usare anche quella logica.
}

