% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{categorie_cup}
\alias{categorie_cup}
\title{Elenco settori, sotto-settori e categorie CUP}
\format{A data frame}
\source{
Team OC
}
\usage{
categorie_cup
}
\description{
Input per la funzione query_cup().
}
\details{
Le variabili sono:
\itemize{
  \item \strong{CUP_COD_SETTORE}: codice settore CUP
  \item CUP_DESCR_SETTORE: ...
  \item \strong{CUP_COD_SOTTOSETTORE}: ...
  \item CUP_DESCR_SOTTOSETTORE: ...
  \item \strong{CUP_COD_CATEGORIA}: ...
  \item CUP_DESCR_CATEGORIA: ...
  \item QUERY: variabile guida per query (inserire 1, 2 o 0)
  \item NOTE: ...
}
}
\keyword{datasets}
