% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{chk_bimestre_sub_fondo_comunitario}
\alias{chk_bimestre_sub_fondo_comunitario}
\title{Workflow chk bimestre - Controllo fondo comunitario}
\usage{
chk_bimestre_sub_fondo_comunitario(progetti, export = FALSE)
}
\arguments{
\item{progetti}{File di tipo progetti da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{export}{Vuoi esportare i report di controllo in TEMP?}
}
\value{
Messaggi e report di controllo.
}
\description{
Workflow di controllo dei dati del nuovo bimestre - Controllo fondo comunitario con focus su YEI (che storicamente ha problemi).
}
