% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{chk_match}
\alias{chk_match}
\title{Verifica sovrapposizione tra due dataset}
\usage{
chk_match(df_left, df_right, id)
}
\arguments{
\item{df_left}{Il dataset a SX.}

\item{df_right}{Il dataset a DX}

\item{id}{La variabile chiave per il join.}
}
\value{
Una matrice di controllo con le seguenti variabili:
 obs: numero di osservazioni totale
 obs_na: numero di obs con id missing (compreso in "obs" e complemento di "obs_n")
 obs_n: numero di obs con id non missing (compreso in "obs")
 obs_m: numero di obs per id molteplici (compreso in "obs_n")
 id_m: numero di id multeplici (compreso in "id_n")
}
\description{
Confronta due dataset per verificare la sorapposizione e gli scarti a sinistra e destra.
}
