% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{chk_nomi_ufficiali}
\alias{chk_nomi_ufficiali}
\title{Verifica allineamento elenco nomi ufficiali DBCOE}
\usage{
chk_nomi_ufficiali(export = TRUE)
}
\arguments{
\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica allineamento elenco nomi ufficiali del DBCOE confrontando file "Dati" e file "Nomi ufficiali".
Controlla sia nomi assenti che nomi presenti ma diversi.
}
