% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{chk_report_psc}
\alias{chk_report_psc}
\title{Controlla report PSC}
\usage{
chk_report_psc(progetti_psc)
}
\arguments{
\item{progetti_psc}{Lista progetti PSC da load_progetti_psc()}
}
\value{
Lista con esito controlli
}
\description{
Controlla allineamento tra report PSC
}
\note{
Il label nella lsita indica la variabile controllata (RISORSE o COE) e i report esaminati nell'ordine della sottrazione.
}
