% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{chk_totali_accordi_fsc}
\alias{chk_totali_accordi_fsc}
\title{Verifica allineamento totali finanziari FSC per accordi coesione}
\usage{
chk_totali_accordi_fsc()
}
\value{
Un messaggio di log.
}
\description{
Verifica allineamento del file "dati" con i totali finanziari FSC per gli accordi per la coesione
}
