% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{chk_variazione_risorse_programmi}
\alias{chk_variazione_risorse_programmi}
\title{Verifica variazione risorse per programma}
\usage{
chk_variazione_risorse_programmi(
  programmi_new = NULL,
  programmi_old = NULL,
  path_to_new = NULL,
  path_to_old = NULL,
  export = FALSE
)
}
\arguments{
\item{programmi_new}{Dataframe da make_report_programmi_coesione()}

\item{programmi_old}{Dataframe da make_report_programmi_coesione()}

\item{path_to_new}{Percorso a csv generato con make_report_programmi_coesione().}

\item{path_to_old}{Percorso a csv generato con make_report_programmi_coesione().}

\item{export}{vuoi salvare il file?}
}
\value{
Un dataframe per programma, ciclo e ambito.
}
\description{
Verifica variazione risorse per programma. Confronta due dataframe risultanti da make_report_programmi_coesione().
}
