% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{crea_accordi_pubblicazione_nofdr}
\alias{crea_accordi_pubblicazione_nofdr}
\title{Crea template pubblicazione accordi per una regione (no FdR)}
\usage{
crea_accordi_pubblicazione_nofdr(regione, interventi)
}
\arguments{
\item{regione}{valore per selezionare AMMINISTRAZIONE_TITOLARE}

\item{interventi}{Elenco interventi da load_db_accordi()}
}
\value{
File in output nella cartella "pubblicazione"
}
\description{
Crea template pubblicazione accordi per una regione - Versione senza FdR
}
