% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{delta_bimestre_sub_sintesi}
\alias{delta_bimestre_sub_sintesi}
\title{Workflow delta bimestre - Controllo per ciclo, ambito e visualizzazione}
\usage{
delta_bimestre_sub_sintesi(progetti, progetti_old)
}
\arguments{
\item{progetti}{File di tipo progetti per il bimestre corrente da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{progetti_old}{File di tipo progetti per un bimestre precedente da load_progetti(light=FALSE, visualizzati=FALSE)}
}
\value{
Messaggi e report di controllo.
}
\description{
Workflow di verifica della variazione dei dati tra due bimestri - Controllo del delta su bimestre precedente, per ciclo, ambito e casistiche di visualizzazione
}
