% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_export.R
\name{export_data}
\alias{export_data}
\title{Export per il dataset finale (DEPRECATA)}
\usage{
export_data(
  pseudo,
  focus,
  bimestre,
  var_ls = NULL,
  var_add = NULL,
  export = TRUE
)
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{focus}{Nome file da salvare in OUTPUT.}

\item{bimestre}{Bimestre di riferimento (utilizzato per la composizone del nome file in OUTPUT.}

\item{var_ls}{Elenco delle variabili di base da esportare. Se nullo usa get_default_vars.}

\item{var_add}{Elenco delle ulteriori variabili da esportare oltre a quelle di var_ls (se si usa get_default_vars)).}

\item{export}{Vuoi salvare?}
}
\value{
Un file "[focus]_[bimestre].csv".
}
\description{
Popola pseudo con una lista di variabili.
}
