% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_export.R
\name{export_data_xls}
\alias{export_data_xls}
\title{Export per il dataset finale in formato excel}
\usage{
export_data_xls(perimetro, focus, bimestre, use_template = FALSE)
}
\arguments{
\item{perimetro}{Dataset "perimetro" creato con export_data().}

\item{focus}{Nome file da salvare in OUTPUT.}

\item{bimestre}{Bimestre di riferimento (utilizzato per la composizone del nome file in OUTPUT.}

\item{use_template}{Vuoi usare un template?}
}
\value{
Un file "[focus]_[bimestre].csv".
}
\description{
Esporta perimetro da export_data() in excel.
}
