% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_export.R
\name{export_sas}
\alias{export_sas}
\title{Export perimetro to SAS}
\usage{
export_sas(
  perimetro,
  focus = "perimetro",
  use_drive = TRUE,
  keep_classe = FALSE,
  split_classe = FALSE
)
}
\arguments{
\item{perimetro}{Dataset "perimetro" da \link[octk]{make_perimetro_edit}o \link[octk]{make_perimetro_std.}}

\item{focus}{Nome del file da salvare.}

\item{use_drive}{Logico. Stai salvano in Drive (TRUE) o in locale (FALSE)?}

\item{keep_classe}{Logico. Vuoi tenere la variabile CLASSE?}

\item{split_classe}{Logico. Vuoi separare N file in base alla variabile CLASSE?}
}
\value{
Un file di tipo "[focus]_clp.csv". Viene salvato in PERIMETRI/OUTPUT_SAS se use_drive == TRUE.
}
\description{
Crea file da passare a SAS per il popolmaneto della variabile FOCUS.
}
