% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{fix_ciclo_cis_taranto_pra_puglia}
\alias{fix_ciclo_cis_taranto_pra_puglia}
\title{Correzione progetti CIS Taranto in PRA Puglia}
\usage{
fix_ciclo_cis_taranto_pra_puglia(progetti_psc)
}
\arguments{
\item{progetti_psc}{Dataset progetti PSC da dentro prep_dati_psc_bimestre()}
}
\value{
dataframe
}
\description{
Correzione progetti CIS Taranto in PRA Puglia, che vengono spostati su ciclo 2014-2020 ma restano nel PRA Puglia
}
\note{
Da usare dentro prep_dati_psc_bimestre()
}
