% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{fix_macroarea_progetti_psc}
\alias{fix_macroarea_progetti_psc}
\title{Correzione macroarea per progetti anomali}
\usage{
fix_macroarea_progetti_psc(progetti_psc)
}
\arguments{
\item{progetti_psc}{Dataset progetti PSC da dentro prep_dati_psc_bimestre()}
}
\value{
dataframe
}
\description{
Correzione macroarea per progetti anomali
}
\note{
Da usare dentro prep_dati_psc_bimestre()
}
