% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{fix_sposta_psc_turismo}
\alias{fix_sposta_psc_turismo}
\title{Associa progetti a PSC Turismo}
\usage{
fix_sposta_psc_turismo(progetti_psc, progetti)
}
\arguments{
\item{progetti_psc}{Dataset progetti PSC da dentro prep_dati_psc_bimestre()}

\item{progetti}{Dataset progetti da load_progetti(bimestre, visualizzati = FALSE, light = FALSE), come usato in dentro prep_dati_psc_bimestre()}
}
\value{
dataframe
}
\description{
Associa progetti a PSC Turismo
}
\note{
Da usare dentro prep_dati_psc_bimestre()
}
