% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{fix_visualizzati_pra_campania}
\alias{fix_visualizzati_pra_campania}
\title{Correzione progetti con OC_FLAG_VISUALIZZAZIONE = 4 in PRA Campania}
\usage{
fix_visualizzati_pra_campania(progetti_psc)
}
\arguments{
\item{progetti_psc}{Dataset progetti PSC da dentro prep_dati_psc_bimestre()}
}
\value{
dataframe
}
\description{
Correzione progetti con OC_FLAG_VISUALIZZAZIONE = 4 in PRA Campania, che sono tutti ripresi perché presenti nel PSC se relativi a compensazioni ambientali e rifiuti (ex OPCM), tranne quello relativo ai debiti di EAV (1MISETPL.EAV01)
}
\note{
Da usare dentro prep_dati_psc_bimestre()
}
