% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_dim_fin}
\alias{get_dim_fin}
\title{Integra classi di dimensione finanziaria}
\usage{
get_dim_fin(df, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}
}
\value{
Il dataset con la variabile CLASSE_FIN, come factor con levels = c("0-100k", "100k-500k", "500k-1M", "1M-2M", "2M-5M", "5M-10M", "10M-infty")
}
\description{
Calcola e integra una variabile con le classi di dimensione finanziaria (x_DIM_FIN).
}
\note{
La modalità **debug** non + implementata.
}
