% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_macroarea}
\alias{get_macroarea}
\title{Definisce la macroarea}
\usage{
get_macroarea(df, progetti, real_reg = TRUE, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti" originale per integrazione.}

\item{real_reg}{Vuoi usare le regioni reali?.}

\item{debug_mode}{Dataset in formato "progetti".}
}
\value{
Il dataset con la variabile x_MACROAREA, come factor con levels = c("Centro-Nord", "Sud", "Trasversale", "Nazionale", "Estero").
}
\description{
Crea la variabile x_MACROAREA.
Le altre denominazioni sono semplificate in "ALTRO TERRITORIO".
}
\note{
La modalità **debug** non + implementata. La modalità **real_reg** forza tutti i progetti
di un programma regionale sul territorio della Regione e sulla relativa macroarea.
}
