% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_macroarea_oc}
\alias{get_macroarea_oc}
\title{Integra la macroarea da PREESTESO}
\usage{
get_macroarea_oc(df, progetti, debug_mode = FALSE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti" in formato PREESTESO per integrazione.}

\item{debug_mode}{Dataset in formato "progetti".}
}
\value{
Il dataset con la variabile x_MACROAREA, come factor con levels = c("Centro-Nord", "Sud", "Trasversale", "Nazionale", "Estero").
}
\description{
Integra la macroarea da PREESTESO in x_MACROAREA partendo da OC_MACROAREA.
}
