% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_regione_simply}
\alias{get_regione_simply}
\title{Semplifica la denominazione di Regioni e Province Autonome}
\usage{
get_regione_simply(df, progetti = NULL, real_reg = TRUE)
}
\arguments{
\item{df}{Dataset con un perimetro in formato "progetti".}

\item{progetti}{Dataset "progetti".}

\item{real_reg}{Vuoi usare le regioni reali?.}
}
\value{
Il dataset con la variabile x_REGIONE, come factor.
}
\description{
Crea la variabile x_REGIONE con la denominazione della Regione o della Provincia Autonoma.
Le altre denominazioni sono semplificate in "ALTRO TERRITORIO".
}
