% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R, R/tools.R
\name{\%!in\%}
\alias{\%!in\%}
\title{!%in%}
\usage{
x \%!in\% y

x \%!in\% y
}
\arguments{
\item{x}{Vettore da verificare.}

\item{y}{Vettore dei casi da ricercare nel vettore da verificare.}
}
\value{
Vettore logico

Vettorelogico con lunghezza di x e valore TRUE per "not in".
}
\description{
Not in

Confronta due vettore per verificare casi "not in".
}
