% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{init_programmazione}
\alias{init_programmazione}
\title{Inizializza il databse della programmazione}
\usage{
init_programmazione(usa_temi = FALSE, export = FALSE)
}
\arguments{
\item{...}{...}
}
\value{
L'intero datbase dei programmazione.
}
\description{
Carica il databse della programmazione, con pulizia della codifica di aree tematiche e temi prioritari FSC.
}
