% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{init_programmazione}
\alias{init_programmazione}
\title{Inizializza il database della programmazione}
\usage{
init_programmazione(
  use_temi = FALSE,
  use_sog = FALSE,
  use_eu = FALSE,
  use_flt = FALSE,
  use_713 = FALSE,
  use_articolaz = FALSE,
  use_location = FALSE,
  use_ciclo = FALSE,
  use_en = FALSE
)
}
\arguments{
\item{usa_temi}{Vuoi caricare il DB con correzione dei temi prioritari?}

\item{usa_sog}{Vuoi caricare il DB con il soggetto programmatore?}

\item{usa_eu}{Vuoi caricare il dataset SIE del DB con le risorse UE e la categoria di regione? (solo per SIE)}

\item{usa_713}{Vuoi caricare anche il DB per il 2007-2013?}
}
\value{
L'intero database dei programmazione, suddiviso in 'po_fesr', 'po_fse', 'po_fsc' e 'po_poc'.
}
\description{
Carica il databse della programmazione, con pulizia della codifica di aree tematiche e temi prioritari FSC.
}
