% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{init_programmazione_dati}
\alias{init_programmazione_dati}
\title{Inizializza il database della programmazione}
\usage{
init_programmazione_dati(
  use_temi = FALSE,
  use_sog = FALSE,
  use_eu = FALSE,
  use_flt = FALSE,
  use_713 = FALSE,
  use_articolaz = FALSE,
  use_location = FALSE,
  use_ciclo = FALSE,
  tipo_ciclo = "CICLO_STRATEGIA",
  use_en = FALSE,
  use_po_psc = FALSE
)
}
\arguments{
\item{use_temi}{Vuoi caricare il DB con correzione dei temi prioritari?}

\item{use_sog}{Vuoi caricare il DB con il soggetto programmatore?}

\item{use_eu}{Vuoi caricare il dataset SIE del DB con le risorse UE e la categoria di regione? (solo per SIE)}

\item{use_713}{Vuoi caricare anche il DB per il 2007-2013?}

\item{tipo_ciclo}{Vuoi usare CICLO_STRATEGIA (default in x_AMBITO nel DB) o CICCLO_RISORSE in senso contabile (sovrascrive x_AMBITO da DB)?}

\item{use_po_psc}{Vuoi usare i dati di programmazione per PO ante art. 44 e non per PSC?}
}
\value{
L'intero database dei programmazione, suddiviso in 'po_fesr', 'po_fse', 'po_fsc' e 'po_poc'.
}
\description{
Carica il databse della programmazione, con pulizia della codifica di aree tematiche e temi prioritari FSC.
}
