% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{init_programmazione_dati}
\alias{init_programmazione_dati}
\title{Inizializza il database della programmazione}
\usage{
init_programmazione_dati(
  DB,
  use_cicli_psc = FALSE,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE
)
}
\arguments{
\item{use_cicli_psc}{Vuoi usare i dati di programmazione per cicli dei PSC?}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni di decisioni in base alle delibere sui POC?}

\item{stime_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le stime di chiusura dei programmi?}
}
\value{
L'intero database dei programmazione, suddiviso in 'po_fesr', 'po_fse', 'po_fsc' e 'po_poc'.
}
\description{
Carica il databse della programmazione, con pulizia della codifica di aree tematiche e temi prioritari FSC.
}
