% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{init_psc}
\alias{init_psc}
\title{Inizializza PSC}
\usage{
init_psc(PSC = NULL)
}
\arguments{
\item{PSC}{Folder di supporto per dati PSC}
}
\value{
Carica i file necessario al flusso nel global environment
}
\description{
Inizializza global environment per flusso di lavoro PSC
}
\note{
...
}
