% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{load_db_accordi_anticipazioni}
\alias{load_db_accordi_anticipazioni}
\title{Carica dati interventi accordi per anticipazioni}
\usage{
load_db_accordi_anticipazioni()
}
\value{
Dataframe
}
\description{
Carica dati interventi accordi per anticipazioni
}
