% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{load_db_accordi_cofinanziamenti}
\alias{load_db_accordi_cofinanziamenti}
\title{Carica dati interventi accordi per cofinanziamenti PR}
\usage{
load_db_accordi_cofinanziamenti(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Dataframe
}
\description{
Carica dati interventi accordi per cofinanziamenti PR
}
