% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{load_db_accordi_complementari}
\alias{load_db_accordi_complementari}
\title{Carica dati interventi accordi per assegnazioni complementari FdR}
\usage{
load_db_accordi_complementari()
}
\value{
Dataframe
}
\description{
Carica dati interventi accordi per assegnazioni complementari FdR
}
