% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{load_db_accordi_completamenti}
\alias{load_db_accordi_completamenti}
\title{Carica dati interventi accordi per completamenti Campania}
\usage{
load_db_accordi_completamenti()
}
\value{
Dataframe
}
\description{
Carica dati interventi accordi per completamenti Campania
}
