% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{load_db_accordi_ordinarie}
\alias{load_db_accordi_ordinarie}
\title{Carica dati interventi accordi per assegnazioni ordinarie}
\usage{
load_db_accordi_ordinarie()
}
\value{
Dataframe
}
\description{
Carica dati interventi accordi per assegnazioni ordinarie
}
