% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{load_db_cis}
\alias{load_db_cis}
\title{Carica dati interventi CIS}
\usage{
load_db_cis(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Dataframe
}
\description{
Carica dati interventi CIS
}
