% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventi_psc.R
\name{load_db_psc}
\alias{load_db_psc}
\title{Carica lista interventi PSC}
\usage{
load_db_psc(DB, use_flt = FALSE)
}
\arguments{
\item{use_flt}{Vuoi caricare solo gli interventi monitorabili (con FLAG_MONITORAGGIO == 1)?}
}
\value{
Dataframe
}
\description{
Carica la lista di interventi delle sezioni ordinarie di PSC dal DBCOE in base alla variabile DB da oc_init().
}
\details{
I progetti privi di OGV rientrano tra gli interventi monitorabili se la delibera di definanziamento non è ancora intervenuta a fronte di istruttoria OGV chiusa.
}
