% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{load_nomi_ufficiali}
\alias{load_nomi_ufficiali}
\title{Carica l'elenco ufficiale dei nomi dal database della programmazione}
\usage{
load_nomi_ufficiali(DB)
}
\arguments{
\item{DB}{Percorso al database programmazione.}
}
\value{
Il dataset con i nomi ufficiali.
}
\description{
Carica l'elenco ufficiale dei nomi dal database della programmazione.
}
