% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{load_operazioni}
\alias{load_operazioni}
\title{Prepara perimetro operazioni con dati coesione}
\usage{
load_operazioni(bimestre, visualizzati = TRUE)
}
\arguments{
\item{visualizzati}{Logico. Vuoi solo i progetti visualizzati sul portale OC?}

\item{perimetro}{Dataset di classe perimetro.}
}
\description{
Prepara perimetro operazioni con dati coesione
}
