% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{load_operazioni}
\alias{load_operazioni}
\title{Prepara perimetro operazioni con dati coesione}
\usage{
load_operazioni(bimestre, usa_meuro = TRUE)
}
\arguments{
\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{perimetro}{Dataset di classe perimetro.}
}
\description{
Prepara perimetro operazioni con dati coesione
}
