% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{load_operazioni_extra}
\alias{load_operazioni_extra}
\title{Carica operazioni 1420}
\usage{
load_operazioni_extra()
}
\value{
Dataset operazinoi da dataiuku.
}
\description{
Carica operazioni 1420 da dataiku
}
