% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{load_progetti}
\alias{load_progetti}
\title{Carica il dataset progetti}
\usage{
load_progetti(bimestre, visualizzati = TRUE, debug = FALSE)
}
\arguments{
\item{bimestre}{Stringa in formato "20180630" come da standard per le date in OC.}

\item{visualizzati}{Logico. Vuoi solo i progetti visualizzati sul portale OC?}

\item{debug}{Logico. Vuoi vedere i totali di progetti e costo pubblico per controllo sul portale OC?}
}
\value{
Il dataset viene caricato come "progetti" nel Global Environment. Se "progetti" è gia presente compare una notifica.
}
\description{
Carica il file progetti_esteso_$BIMESTRE.csv dal folder DATI.
}
