% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{load_progetti_psc}
\alias{load_progetti_psc}
\title{Carica dati di base per PSC}
\usage{
load_progetti_psc(bimemstre)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}
}
\value{
Dataframe
}
\description{
Carica dati di base per PSC
}
