% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{load_progetti_psc}
\alias{load_progetti_psc}
\title{Carica dati di base per PSC}
\usage{
load_progetti_psc(bimestre, fix_no_temi_no_coe = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{fix_no_temi_no_coe}{Logico. Vuoi scartare i progetti con tema missing e finanziamenti pari a 0?}

\item{versione}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}
}
\value{
Dataframe
}
\description{
Carica dati di base per PSC
}
