% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{load_stime_siepoc1420}
\alias{load_stime_siepoc1420}
\title{Carica un dataset "stime" dal database della programmazione}
\usage{
load_stime_siepoc1420(DB)
}
\arguments{
\item{DB}{Percorso al database generato con oc_init() o sovrascritto.}
}
\value{
Il dataset "stime".
}
\description{
Carica il dataset "stime" per SIE e POC 2014-2020 richiesto dal database della programmazione.
}
