% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_classi.R
\name{make_classi}
\alias{make_classi}
\title{Classificazione per categoria CUP e campo d'intervento UE}
\usage{
make_classi(
  pseudo,
  classe_jolly = "Altro",
  livelli_classe = NULL,
  export = TRUE,
  debug = FALSE
)
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{classe_jolly}{Nome per classe da assegnare ai casi non mappati in classi_cup e classi_ue}

\item{livelli_classe}{Nomi per factor di classe}

\item{export}{Logico. Vuoi salvare?}

\item{debug}{Logico. Vuoi una matrice con freq cup x ue}
}
\value{
Un file "pseudo".
}
\description{
Classificazione per categoria CUP e campo d'intervento UE
}
