% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_classi.R
\name{make_classi_hard_soft}
\alias{make_classi_hard_soft}
\title{Classificazione Hard-Soft}
\usage{
make_classi_hard_soft(pseudo, export = TRUE)
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{export}{Logico. Vuoi salvare?}
}
\value{
Un file "pseudo".
}
\description{
Classificazione Hard-Soft da nautra CUP
}
