% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{make_input_delta}
\alias{make_input_delta}
\title{Calcola delta da precedente input}
\usage{
make_input_delta(OLD)
}
\arguments{
\item{OLD}{Percorso alla cartella di livello WORK della precedente elaborazione presa a riferimento.}
}
\value{
Calcola la differenza tra il file indicato in WORK rispetto ai template nel package, salvando un file "input_query_delta.xlsx" da editare manualmente.
}
\description{
Calcola delta da precedente input in format su file excel.
}
