% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_opendata_dotazioni}
\alias{make_opendata_dotazioni}
\title{Opendata per dotazioni}
\usage{
make_opendata_dotazioni(
  programmi = NULL,
  progetti = NULL,
  use_fix_siepoc = TRUE,
  stime_fix_siepoc = FALSE,
  export = TRUE,
  export_xls = TRUE
)
}
\arguments{
\item{programmi}{Dati di base da workflow_programmazione().}

\item{progetti}{Dataset di tipo 'progetti' (serve per denominazioni programmi da sito e non da DB)}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni?}

\item{stime_fix_siepoc}{Per correggere i dati SIE e POC 1420 con le anticipazioni vuoi usare anche le stime?}

\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx per ciclo e ambito in OUTPUT?}
}
\value{
File opendata con le dotazioni per ambito e per i cicli 2007-2013 e 2014-2020.
}
\description{
Crea il file opendata con le dotazioni dei programmi da pubblicare sul sito di OC
}
\note{
...
}
