% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_pagina_programmi}
\alias{make_pagina_programmi}
\title{Lista programmi per pagina dedicata}
\usage{
make_pagina_programmi(
  programmi = NULL,
  progetti = NULL,
  use_fix_siepoc = FALSE,
  stime_fix_siepoc = FALSE,
  export = TRUE
)
}
\arguments{
\item{programmi}{Dati di base da workflow_programmazione().}

\item{progetti}{Dataset di tipo 'operazioni' (serve per verificare i programmi pubblicati)}

\item{use_fix_siepoc}{Vuoi correggere i dati SIE e POC 1420 con le anticipazioni?}

\item{stime_fix_siepoc}{Per correggere i dati SIE e POC 1420 con le anticipazioni vuoi usare anche le stime?}

\item{export}{Vuoi salvare il file?}
}
\value{
Lista dei programmi 2007-2013 e 2014-2020 applicando le convenzioni per la pubblicazione nella pagina "programmi" del sito.
}
\description{
Crea la lista dei programmi da pubblicare nella "pagina programmi" del sito di OC
}
