% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_pagina_programmi}
\alias{make_pagina_programmi}
\title{Lista programmi per pagina dedicata}
\usage{
make_pagina_programmi(programmi = NULL, progetti = NULL, export = TRUE)
}
\arguments{
\item{programmi}{Dati di base da workflow_programmazione().}

\item{progetti}{Dataset di tipo 'progetti' (serve per denominazioni programmi da sito e non da DB)}

\item{export}{vuoi salvare il file?}
}
\value{
Lista dei programmi 2007-2013 e 2014-2020
}
\description{
Crea la lista dei programmi da pubblicare nella "pagina programmi" del sito di OC
}
