% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_perimetro.R
\name{make_perimetro_std}
\alias{make_perimetro_std}
\title{Definizione del perimetro (versione standard per Turismo)}
\usage{
make_perimetro_std(pseudo, export = TRUE, stoplist = NULL,
  safelist = NULL, debug = FALSE, progetti = NULL, var_ls = NULL)
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{export}{Vuoi salvare pseduo.csv in TEMP?}

\item{stoplist}{Dataset con l'elenco dei COD_LOCALE_PROGETTO da eliminare.}

\item{safelist}{Dataset con l'elenco dei COD_LOCALE_PROGETTO da conservare anche se sarebbero scartati.}

\item{debug}{Vuoi salvare scarti.csv in TEMP, con i progetti non considerati?}

\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}

\item{var_ls}{Varibili da integrare in caso di debug.}
}
\value{
Un dataframe "pseudo" con le variabili addizionali "CHK" e "PERI"
#' @section Warning:
Al momento è disponibile solo la versione con QUERY_CUP, QUERY_PO e QUERY_UE.
}
\description{
Ricerca progetti per settore, sotto-settore e categoria CUP a partire da input in "categorie_cup.csv".
}
