% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{make_po_riclass}
\alias{make_po_riclass}
\title{Crea nuovo po_riclass.csv}
\usage{
make_po_riclass(bimestre, progetti = NULL)
}
\arguments{
\item{bimestre}{Bimestre di riferimento.}

\item{progetti}{Dataset completo da progetti_esteso.csv.}
}
\value{
Il dataset viene salvato in DATA e può essere caricato con load_progetti(light = TRUE).
}
\description{
DO SOMETHING.
}
