% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{make_pseudo_edit}
\alias{make_pseudo_edit}
\title{Wrapper per query editabile}
\usage{
make_pseudo_edit(progetti, query_ls = c("query_cup"), export = TRUE)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_[1], QUERY_[2], QUERY_[N] e TIPO_QUERY.
}
\description{
Wrapper per esecuzione di una lista di query a composizione libera.
}
\section{Warning}{

I dati valori NA sono convertiti in 0.
La valutazione dinamica di TIPO_QUERY è ancora da implementare.
}

