% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{make_report_macroaree_coesione}
\alias{make_report_macroaree_coesione}
\title{Esporta report per ciclo, ambito e macroarea con dati coesione}
\usage{
make_report_macroaree_coesione(programmi, usa_meuro = TRUE, export = TRUE)
}
\arguments{
\item{programmi}{Dataset in formato "programmi" da make_report_programmi_macroaree_coesione()}

\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro. Attenzione: per usare Meuro il perimetro deve essere in euro, viene arrotondato dopo}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per ciclo, ambto e macroarea.
}
\description{
Report con apertura ciclo, ambito e macroarea.
E' costruito su operazioni e dati coesione.
}
