% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{make_report_programmi_coesione}
\alias{make_report_programmi_coesione}
\title{Esporta report per Programmi con dati coesione}
\usage{
make_report_programmi_coesione(perimetro, usa_meuro = FALSE,
  add_713 = FALSE, add_totali = FALSE, focus = "report",
  export = FALSE)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{add_713}{Vuoi caricare anche i dati di programmaizone per il 2007-2013?}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per programma e fase procedurale.
}
\description{
Report con apertura per programma e fase procedurale rispetto al focus selezionato.
E' costruito su operaizoni e dati coesione.
}
