% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{make_report_programmi_coesione}
\alias{make_report_programmi_coesione}
\title{Esporta report per Programmi con dati coesione}
\usage{
make_report_programmi_coesione(
  perimetro,
  usa_meuro = FALSE,
  use_713 = FALSE,
  add_totali = FALSE,
  use_cp2 = FALSE,
  focus = "report",
  export = FALSE,
  progetti = NULL,
  po_riclass = NULL
)
}
\arguments{
\item{perimetro}{Dataset di classe perimetro.}

\item{usa_meuro}{Vuoi i dati in Meuro? Di default sono in euro. Attenzione: per usare Meuro il perimetro deve essere in euro, viene arrotondato dopo}

\item{use_713}{Vuoi caricare anche i dati di programmaizone per il 2007-2013?}

\item{add_totali}{Vuoi aggiungere valori calcolati in termini di costo pubblico?}

\item{use_cp2}{Se add_totali == TRUE, vuoi raddoppiare i valori relativi ai progetti multi-programma?}

\item{focus}{nome per file.}

\item{export}{vuoi salvare il file?}

\item{progetti}{dataset di tipo "progetti" da utilizzare per con add_totali == TRUE}

\item{po_riclass}{dataset di tipo "po_riclass" da utilizzare (altrimenti usa default nel package)}
}
\value{
Un file csv con apertura per programma e fase procedurale.
}
\description{
Report con apertura per programma e fase procedurale rispetto al focus selezionato.
E' costruito su operaizoni e dati coesione.
}
