% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_report_risorse}
\alias{make_report_risorse}
\title{Esporta report con risorse coesione per ciclo, ambito e macroarea}
\usage{
make_report_risorse(
  ciclo = NULL,
  use_meuro = FALSE,
  use_flt = FALSE,
  tipo_ciclo = "CICLO_STRATEGIA",
  export = FALSE
)
}
\arguments{
\item{ciclo}{Dataset di classe perimetro.}

\item{use_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{tipo_ciclo}{Vuoi usare CICLO_STRATEGIA (default in x_AMBITO nel DB) o CICCLO_RISORSE in senso contabile (sovrascrive x_AMBITO da DB)?}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per ciclo e macroarea.
}
\description{
Esporta report con risorse coesione per ciclo, ambito e macroarea.
}
