% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{make_report_risorse}
\alias{make_report_risorse}
\title{Esporta report con risorse coesione per ciclo, ambito e macroarea}
\usage{
make_report_risorse(
  ciclo = NULL,
  use_meuro = FALSE,
  use_flt = FALSE,
  use_eu = FALSE,
  use_po_psc = FALSE,
  force_yei = FALSE,
  tipo_ciclo = "CICLO_STRATEGIA",
  export = FALSE
)
}
\arguments{
\item{ciclo}{Vuoi un ciclo specifico?}

\item{use_meuro}{Vuoi i dati in Meuro? Di default sono in euro.}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{use_eu}{Logico. vuoi vedere risorse UE ove previste?}

\item{use_po_psc}{Vuoi usare i dati di programmazione per PO ante art. 44 e non per PSC?}

\item{force_yei}{Logico. Vuoi forzare FSE in YEI?}

\item{tipo_ciclo}{Vuoi usare CICLO_STRATEGIA (default in x_AMBITO nel DB) o CICCLO_RISORSE in senso contabile (sovrascrive x_AMBITO da DB)?}

\item{export}{vuoi salvare il file?}
}
\value{
Un file csv con apertura per ciclo e macroarea.
}
\description{
Esporta report con risorse coesione per ciclo, ambito e macroarea.
}
