% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{make_report_temi_stato_psc}
\alias{make_report_temi_stato_psc}
\title{Report PSC per temi con stato procedurale}
\usage{
make_report_temi_stato_psc(
  progetti_psc,
  programmazione = NULL,
  visualizzati = TRUE,
  usa_meuro = FALSE,
  export = FALSE,
  export_xls = FALSE
)
}
\arguments{
\item{progetti_psc}{Dataset da load_progetti_psc()}

\item{programmazione}{Dati di programmazione DBCOE di tipo "fsc_matrice_po_psc.xlsx"}

\item{visualizzati}{Logico. Vuoi solo i progetti visualizzati sul portale OC?}

\item{usa_meuro}{Logico. Vuoi dati in Meuro?}

\item{export}{Vuoi salvare il file csv in TEMP?}

\item{export_xls}{Vuoi salvare i file xlsx per ciclo e ambito in OUTPUT?}
}
\value{
Report di confronto programmazione attuazione per PSC e PO in essi confluiti. I nuovi
}
\description{
Crea report di confronto programmazione attuazione per temi con apertura per stato procedurale
}
\note{
...
}
