% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{matrix_opos}
\alias{matrix_opos}
\title{Obiettivi Specifici 2021-2027}
\format{
A data frame
}
\source{
Team OC su regolamenti in corso di negoziazione
}
\usage{
matrix_opos
}
\description{
... OPOS
}
\details{
Le variabili sono:
\itemize{
  \item FONDO: (CHK)
  \item OP: (CHK)
  \item \strong{OPOS}: codice obiettivo specifico (TODO: da riformulare)
  \item OS_DESCR_EN: ...
  \item OS_DESCR: ... (create con google translate)
  \item OS_DESCR_SHORT: nome sintetico del'OPOS
}
}
\keyword{datasets}
