% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{matrix_ra_temi_fsc}
\alias{matrix_ra_temi_fsc}
\title{Raccordo tra RA e temi prioritari FSC 2014-2020}
\format{A data frame}
\source{
Team OC
}
\usage{
matrix_ra_temi_fsc
}
\description{
... RA vs temi FSC
}
\details{
Le variabili sono:
\itemize{
  \item \strong{COD_RA}: codice risultato atteso
  \item RA_ID: ...
  \item RA: ... (da rinominare in RA_DESC)
  \item \strong{COD_SETTORE_STRATEGICO}: area tematica FSC (denominazione BDU?)
  \item x_AREA_TEMATICA: denominazione area tematica FSC
  \item \strong{COD_ASSE_TEMATICO}: tema prioritario FSC (denominazione BDU?)
  \item x_TEMA_PRIORITARIO: denominazione tema prioritario FSC
  \item NOTE: ...
}
}
\keyword{datasets}
