% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{monithon_clp}
\alias{monithon_clp}
\title{Report su Monithon}
\format{
A data frame
}
\source{
Team OC su dati \url{http://monithon.it}
}
\usage{
monithon_clp
}
\description{
Riconciliazione tra progetti di OpenCoesione e report di monitoraggio civico su Monithon.
}
\section{Warning}{

Dati aggiornati ad aprile 2018
}

\keyword{datasets}
