% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{oc_init}
\alias{oc_init}
\title{Inizializza OCTK}
\usage{
oc_init(
  bimestre,
  db_ver,
  data_path = NULL,
  db_path = NULL,
  workarea = NULL,
  elab = NULL,
  focus = NULL,
  ver = NULL,
  use_drive = TRUE,
  drive_root = NULL,
  DEV_MODE = FALSE
)
}
\arguments{
\item{bimestre}{Versione dei dati di attuazione da utilizzare. Stringa in formato "20180630" come da standard per le date in OC.}

\item{db_ver}{Versione dei dati di programmazione da utilizzare. Stringa in formato "20180630" come da standard per le date in OC.}

\item{data_path}{Percorso allla fonte dati.}

\item{db_path}{Percorso al bd di programmazione.}

\item{workarea}{Percorso per la cartella di lavoro. Se vuoto, va di default nella wd.}

\item{elab}{Nome della cartella dedicata al tipo di elebaorazione (solo con use_drive=TRUE)}

\item{focus}{Nome della sotto-cartella dedicata all'elebaorazione (solo con use_drive=TRUE). Il nome viene usato anche nel naming dei file in output.}

\item{ver}{Nome della sotto-sotto-cartella dedicata alla versione dell'elebaorazione (solo con use_drive=TRUE).}

\item{use_drive}{Logico. Vuoi usare GoogleDrive?}

\item{drive_root}{Percorso per la cartella di lavoro su GoogleDrive.}

\item{DEV_MODE}{Logico. Vuoi usare la DEV_MODE?}
}
\value{
I path della workarea (WORK, con INPUT, OUTPUT e TEMP), dei dati di attuazione (DATA)
e del database di programmazione (DB) sono disponibili nel Global Environment. Se non esistono, vengono creati i folder
"input", "output" e "temp" nella workarea di progetto denominata come "focus" all'interno di "...".
}
\description{
Definisce i puntamenti ai folder di lavoro e ai dati.
}
