% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{po_linee_azioni}
\alias{po_linee_azioni}
\title{Elenco programmi, linee e azioni}
\format{
A data frame
}
\source{
Team OC su dati BDU
}
\usage{
po_linee_azioni
}
\description{
Input per la funzione query_po().
Creata con make_matrix_po() in setup_data.R.
La logica segue quella in due livelli per articolazioni e sub-articolazioni in "progetti esteso".
Ad esempio: "Asse" e "Obiettivo operativo" per i PO FS 2007-2013.
}
\details{
Le variabili sono:
\itemize{
  \item OC_COD_CICLO: ...
  \item OC_DESCR_CICLO: ...
  \item \strong{OC_CODICE_PROGRAMMA}: codice settore CUP
  \item OC_DESCRIZIONE_PROGRAMMA: ...
  \item OC_ARTICOLAZIONE_PROGRAMMA: ...
  \item \strong{OC_COD_ARTICOLAZ_PROGRAMMA}: ...
  \item OC_DESCR_ARTICOLAZ_PROGRAMMA: ...
  \item OC_SUBARTICOLAZIONE_PROGRAMMA: ...
  \item \strong{OC_COD_SUBARTICOLAZ_PROGRAMMA}: ...
  \item OC_DESCR_SUBARTICOLAZ_PROGRAMMA: ...
  \item QUERY: variabile guida per query (inserire 1, 2 o 0)
  \item NOTE: ...
}
}
\section{Warning}{

Dati da aggiornare per ogni bimestre.
}

\keyword{datasets}
