% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{po_riclass}
\alias{po_riclass}
\title{Riclassifica i programmi}
\format{A data frame}
\source{
Team OC
}
\usage{
po_riclass
}
\description{
Asset con la riclassificazione dei programmi per il miglioramento della reportistica.
Contiene correzioni puntuali forzate rispetto ai dati originali di OC.
}
\details{
Le variabili sono:
\itemize{
  \item \strong{OC_CODICE_PROGRAMMA}: codice programma (per match con i dati OC)
  \item x_CICLO: ciclo
  \item x_FONDO: ...
  \item x_GRUPPO: ...
  \item x_AMBITO: ... (TODO!!!)
  \item x_PROGRAMMA: nome sintetico del programma
}
}
\keyword{datasets}
