% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{prep_dati_psc_bimestre}
\alias{prep_dati_psc_bimestre}
\title{Dati FSC per elaborazioni PSC}
\usage{
prep_dati_psc_bimestre(
  bimestre,
  matrix_po_psc,
  po_naz,
  art44,
  matrix_1420,
  matrix_713,
  matrix_temi_settori,
  progetti = NULL,
  progetti_pub = NULL
)
}
\arguments{
\item{bimestre}{Bimestre di riferimento, nella versione specifica per i dati PSC (es. 20211231.01 oppure 20211231), che viene ricondotto automaticamente al bimestre OC}

\item{matrix_po_psc}{Matrice di riconciliazione PO - PSC}

\item{art44}{Dati di base per istruttoria art44 al 20200630}

\item{matrix_1420}{Matrice di riconciliazione tra temi prioritari FSC 2014-2020 e settori di intervento PSC}

\item{matrix_713}{Matrice di riconciliazione tra obiettivi specifici QSN 2007-2013 e settori di intervento PSC}

\item{matrix_temi_settori}{Dominio per aree tematiche e settori di intervento PSC}

\item{progetti}{Dataset progetti da load_progetti(visualizzati = FALSE, light = FALSE)}

\item{progetti_pub}{Dataset progetti esteso pubblicato}
}
\value{
File "dati_psc_BIMESTRE.csv" in TEMP
}
\description{
Crea il file con operazioni FSC compatibile con PSC
}
\note{
...
}
