% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_migrati.R
\name{prep_dati_psc_migrati_bimestre}
\alias{prep_dati_psc_migrati_bimestre}
\title{Dati FSC per elaborazioni PSC}
\usage{
prep_dati_psc_migrati_bimestre(
  bimestre,
  versione_psc,
  operazioni,
  progetti,
  chk_today,
  export = TRUE
)
}
\arguments{
\item{bimestre}{Bimestre di riferimento, nella versione specifica per i dati PSC (es. 20211231.01 oppure 20211231), che viene ricondotto automaticamente al bimestre OC}

\item{versione_psc}{Versione di riferimento dei dati (sono possibili più versioni per lo stesso bimestre)}

\item{operazioni}{Dataset progetti da load_operazioni(visualizzati = FALSE)}

\item{progetti}{Dataset progetti da load_progetti(visualizzati = FALSE, light = FALSE)}

\item{chk_today}{Parametro da passare a get_stato_attuazione(), con formato "2021-02-28"}

\item{export}{Vuoi esportare i dati nel repository PSC in DATI?}
}
\value{
File "dati_psc_BIMESTRE_VERSIONE.csv" in OUTPUT
}
\description{
Crea il file con operazioni FSC compatibile con PSC
}
\note{
...
}
