% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{query_atsi_psc}
\alias{query_atsi_psc}
\title{Ricerca progetti per aree tematiche e settori di intervento PSC}
\usage{
query_atsi_psc(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_ATSI_PSC.
}
\description{
Ricerca progetti per aree tematiche e settori di intervento PSC (ATSI) a partire da input in "aree_temi_psc.csv".
}
