% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_po}
\alias{query_po}
\title{Ricerca progetti per articolazione}
\usage{
query_po(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_PO.
}
\description{
Ricerca progetti per programma, linea (articolazione) e azione (sub-articolazione) a partire da input in "po_linee_azioni.csv".
}
