% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{query_progcomp}
\alias{query_progcomp}
\title{Ricerca progetti per progetto complesso}
\usage{
query_progcomp(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_PROGCOMP.
}
\description{
Ricerca progetti per progetto complesso a partire da input in "prog_comp.csv".
}
