% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{query_qsn}
\alias{query_qsn}
\title{Ricerca progetti per obiettivi QSN}
\usage{
query_qsn(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_QSN.
}
\description{
Ricerca progetti per obiettivi generali e specifici QSN 2007-2013 a partire da input in "qsn".
}
