% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_ra}
\alias{query_ra}
\title{Ricerca progetti per Risultato atteso}
\usage{
query_ra(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_RA.
}
\description{
Ricerca progetti per Risultato atteso (RA) a partire da input in "ra.csv".
}
