% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_strum}
\alias{query_strum}
\title{Ricerca progetti per strumento attuativo}
\usage{
query_strum(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_STRUM.
}
\description{
Ricerca progetti per strumento attuativo a partire da input in "strum_att.csv".
}
