% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{query_tipo_cup}
\alias{query_tipo_cup}
\title{Ricerca progetti per tipologia CUP}
\usage{
query_tipo_cup(progetti)
}
\arguments{
\item{progetti}{Dataset "progetti_esteso_<BIMESTRE>.csv".}
}
\value{
Un dataframe con COD_LOCALE_PROGETTO, QUERY_TIPO_CUP.
}
\description{
Ricerca progetti per natura e tipologia CUP a partire da input in "tipologie_cup".
}
