% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{ra}
\alias{ra}
\title{Elenco risultati attesi AP 2014-2020}
\format{
A data frame
}
\source{
Team OC
}
\usage{
ra
}
\description{
Input per la funzione ... (TODO).
}
\details{
Le variabili sono:
\itemize{
  \item OC_COD_CICLO: ...
  \item \strong{COD_RA}: codice del risultato atteso 2014-2020
  \item RA_DESC: ...
  \item QUERY: variabile guida per query (inserire 1, 2 o 0)
  \item NOTE: ...
}

#' @section Warning:
Da integrare con RA rural (CHK?).
}
\keyword{datasets}
