% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_migrati.R
\name{ranking_data_quality_psc}
\alias{ranking_data_quality_psc}
\title{Ranking qualità dati PSC}
\usage{
ranking_data_quality_psc(report, export = TRUE)
}
\arguments{
\item{report}{Report di analisi qualità dati da report_data_quality_psc()}

\item{export}{Vuoi esportare il file in OUTPUT?}
}
\value{
Dataframe
}
\description{
Ranking qualità dati PSC
}
