% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{refactor_ambito}
\alias{refactor_ambito}
\title{Refactor di x_AMBITO}
\usage{
refactor_ambito(df)
}
\arguments{
\item{df}{Dataset progetti in formato standard.}
}
\value{
Il dataset integrato.
}
\description{
Integra un perimetro di progetti con factor di x_AMBITO.
}
