% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_migrati.R
\name{report_data_quality_psc}
\alias{report_data_quality_psc}
\title{Report qualità dati PSC}
\usage{
report_data_quality_psc(analisi, progetti_psc, export = TRUE)
}
\arguments{
\item{analisi}{Dataset di analisi qualità dati da analisi_data_quality_psc()}

\item{progetti_psc}{Dataset da load_progetti_psc_migrati()}

\item{export}{Vuoi esportare il file in OUTPUT?}
}
\value{
Dataframe
}
\description{
Report qualità dati PSC
}
