% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{ricodifica_macroaree}
\alias{ricodifica_macroaree}
\title{Ricodifica la voce macroarea lato programmazione come x_MACROAREA}
\usage{
ricodifica_macroaree(programmi)
}
\arguments{
\item{programmi}{Dataframe da init_programmazione_dati()}
}
\value{
Un dataframe con x_MACROAREA
}
\description{
Ricodifica la voce macroarea lato programmazione come x_MACROAREA
}
