% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classi.R
\name{setup_classi_cup}
\alias{setup_classi_cup}
\title{Crea nuova base per riclassificazione lato categorie CUP}
\usage{
setup_classi_cup(pseudo, file_name = "classi_cup_NEW.csv")
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{file_name}{Nome file da salvare in INPUT.}
}
\value{
Un file "classi_cup.csv".
#' @section Warning:
Da rinominare in "classi_cup.csv" e modificare.
}
\description{
...
}
