% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_classi.R
\name{setup_classi_ue}
\alias{setup_classi_ue}
\title{Crea nuova base per riclassificazione lato temi/campi UE}
\usage{
setup_classi_ue(pseudo, file_name = "classi_ue_NEW.csv")
}
\arguments{
\item{pseudo}{Dataset "pseudo".}

\item{file_name}{Nome file da salvare in INPUT.}
}
\value{
Un file "classi_ue.csv".
#' @section Warning:
Da rinominare in "classi_ue.csv" e modificare.
}
\description{
...
}
