% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{setup_macroaree_sie}
\alias{setup_macroaree_sie}
\title{Crea file di base per macroaree}
\usage{
setup_macroaree_sie(bimestre, operazioni, export = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{export}{vuoi salvare il file?}

\item{perimetro}{Dataset di classe operazioni}
}
\value{
Dataset macroaree per i report make_report_programmi_macroaree e make_report_macroaree_coesione
}
\description{
Crea file di base per macroaree, da operazioni.
}
