% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{setup_operazioni}
\alias{setup_operazioni}
\title{Crea il dataset operazioni}
\usage{
setup_operazioni(
  bimestre,
  progetti,
  operazioni_713,
  operazioni_1420,
  operazioni_extra,
  export = FALSE,
  use_fix = FALSE,
  use_ecomix = FALSE,
  use_sito = FALSE,
  debug = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{progetti}{Dataset con un perimetro in formato "progetti".}

\item{operazioni_713}{File di tipo operazioni da flusso sas/dataiku.}

\item{operazioni_1420}{File di tipo operazioni da flusso sas/dataiku.}

\item{operazioni_extra}{File di tipo operazioni da flusso sas/dataiku.}

\item{use_fix}{Vuoi applicare fix_progetti() prima di esecuzione? Non impatta su x_AMBITO!}

\item{use_ecomix}{Vuoi calcolare il costo coesione al netto delle economie solo per i progetti conclusi?}

\item{use_sito}{Vuoi usare il calcolo dellle risorse coesione fatto per il sito OC?}
}
\value{
Il dataset operazioni con le variabili coesione calcolate: COE, COE_IMP e COE_PAG.
}
\description{
Crea il dataset operazioni a partire da sottoprodotti del workflow SAS.
}
\note{
La modalità **debug** esporta diversi csv in TEMP La modalità **export** esporta operazioni_light.csv in DATA.
}
