% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coesione.R
\name{setup_operazioni}
\alias{setup_operazioni}
\title{Crea il dataset operazioni}
\usage{
setup_operazioni(bimestre, progetti = NULL, export = FALSE,
  use_fix = FALSE, debug = FALSE)
}
\arguments{
\item{bimestre}{Bimestre di riferimento}

\item{progetti}{Dataset con un perimetro in formato "progetti".}
}
\value{
Il dataset operazioni con le variabili coesione calcolate: COE, COE_IMP e COE_PAG.
}
\description{
Crea il dataset operazioni a partire da sottoprodotti del workflow SAS.
}
\note{
La modalità **debug** esporta diversi csv in TEMP La modalità **export** esporta operazioni_light.csv in DATA.
}
