% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{setup_query}
\alias{setup_query}
\title{Setup query e stoplist}
\usage{
setup_query()
}
\value{
File csv in INPUT
}
\description{
Salva i file csv del blocco "query" in INPUT insieme a stoplist.csv, safelist.csv e fixlist.csv
}
