% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_query.R
\name{setup_query_xls}
\alias{setup_query_xls}
\title{Setup query e stoplist in excel}
\usage{
setup_query_xls()
}
\value{
File xls e csv in INPUt
}
\description{
Salva in INPUT un file xls con i file csv del blocco "query", insieme a stoplist.csv, safelist.csv e fixlist.csv
}
