% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{workflow_accordi_monitoraggio}
\alias{workflow_accordi_monitoraggio}
\title{Workflow per template monitoraggio extra-sistema accordi}
\usage{
workflow_accordi_monitoraggio(interventi, template)
}
\arguments{
\item{interventi}{Elenco interventi da load_db_accordi()}

\item{template}{Nome del file xlsx in input}
}
\value{
File in output nella cartella "monitoraggio"
}
\description{
Workflow per template monitoraggio extra-sistema accordi
}
