% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordi.R
\name{workflow_accordi_pubblicazione}
\alias{workflow_accordi_pubblicazione}
\title{Workflow per template pubblicazione accordi}
\usage{
workflow_accordi_pubblicazione(interventi)
}
\arguments{
\item{interventi}{Elenco interventi da load_db_accordi()}
}
\value{
File in output nella cartella "pubblicazione"
}
\description{
Workflow per template pubblicazione accordi
}
