% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{workflow_cis}
\alias{workflow_cis}
\title{Workflow per analisi CIS}
\usage{
workflow_cis(
  bimestre,
  interventi = NULL,
  operazioni = NULL,
  progetti = NULL,
  debug = FALSE,
  debug_peri = FALSE,
  export = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di OpenCoesione}

\item{interventi}{Interventi CIS da load_db_interventi(tipo = "CIS)}

\item{operazioni}{Dataset da operazioni_light}

\item{progetti}{Dataset da progetti_light}

\item{debug}{Vuoi verificare il DB e il mappging con i clp?}

\item{debug_peri}{Vuoi verificare i se tutti i clp nel perimetro sono mappati nel DB?}

\item{export}{Vuoi esportare in TEMP in xlsx?}
}
\description{
Workflow per analisi CIS
}
