% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tools.R
\name{workflow_delta_bimestre}
\alias{workflow_delta_bimestre}
\title{Workflow delta bimestre}
\usage{
workflow_delta_bimestre(progetti, progetti_old)
}
\arguments{
\item{progetti}{File di tipo progetti per il bimestre corrente da load_progetti(light=FALSE, visualizzati=FALSE)}

\item{progetti_old}{File di tipo progetti per un bimestre precedente da load_progetti(light=FALSE, visualizzati=FALSE)}
}
\value{
Messaggi e report di controllo.
}
\description{
Workflow di verifica della variazione dei dati tra due bimestri
}
