% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroaree.R
\name{workflow_macroaree}
\alias{workflow_macroaree}
\title{Workflow per creare il dataset macroaree}
\usage{
workflow_macroaree(
  bimestre,
  progetti,
  operazioni_713,
  operazioni_1420,
  operazioni_extra,
  riprop_yei = FALSE,
  debug = FALSE
)
}
\arguments{
\item{bimestre}{Bimestre di rifeirimento da oc_init().}

\item{progetti}{Dataset "progetti" in formato PREESTESO per integrazione.}

\item{operazioni_713}{File di tipo operazioni da flusso sas/dataiku.}

\item{operazioni_1420}{File di tipo operazioni da flusso sas/dataiku.}

\item{riprop_yei}{Vuoi riproporzionare yei dove mancano dati finanziari per livelli gerarchici? Default su FALSE.}

\item{debug_mode}{Esporta i file relativi a [...]}
}
\value{
Il dataset con la variabile x_MACROAREA, da usare all'interno di setup_macroare().
}
\description{
Workflow per creare il dataset macroaree
}
