% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmazione.R
\name{workflow_programmazione}
\alias{workflow_programmazione}
\title{Workflow per utilizzo dataset della programmazione.}
\usage{
workflow_programmazione(use_info = FALSE, use_flt = TRUE, progetti = NULL)
}
\arguments{
\item{use_info}{Logico. Vuoi aggiungere le informazioni di supporto?}

\item{use_flt}{Logico. Vuoi utilizzare solo i programmi che rientrano nel perimetro coesione monitorabile?}

\item{progetti}{Dataset progetti importato con load_progetti}
}
\value{
Il dataset dei programmi con risorse e evewntualmente infomrmazioni di supporto
}
\description{
Workflow per utilizzo dataset della programmazione.
}
